### 2026-02-17 #752 @mauroloprete
metasurvey — Reproducible Survey Data Processing with Step Pipelines
Submitting Author Name: Mauro Loprete
Submitting Author Github Handle: @mauroloprete
Repository: https://github.com/metasurveyr/metasurvey
Version submitted: 0.0.21
Submission type: Standard

Version accepted: TBD

---

```
Package: metasurvey
Title: Reproducible Survey Data Processing with Step Pipelines
Version: 0.0.21
    c(
        person(
            given = "Mauro",
            family = "Loprete",
            email = "mauro.loprete@icloud.com",
            comment = c(ORCID = "0000-0003-1560-0183")
        ),
        person(
            given = "Natalia",
            family = "da Silva",
            role = c("aut"),
            email = "natalia.dasilva@fcea.edu.uy",
            comment = c(ORCID = "0000-0002-6031-7451")
        ),
        person(
            given = "Fabricio",
            family = "Machado",
            role = c("aut"),
            email = "fabricio.mch.slv@gmail.com"
        )
    )
Description: Wraps the 'survey' package with 'R6' classes and a step-based
    pipeline for reproducible survey data processing. Supports complex sampling
    designs including rotating panels with bootstrap replicate weights, and
    provides a recipe system for sharing and reproducing data transformation
    workflows across survey editions.
```

- Please indicate which category or categories from our [package fit policies](https://devguide.ropensci.org/softwarereviewpolicies.html#package-categories) this package falls under:

- Explain how and why the package falls under these categories (1-2 paragraphs):

  **Data munging**: metasurvey provides a lazy step pipeline (`step_compute`, `step_recode`, `step_rename`, `step_remove`, `step_join`, `step_filter`) where every transformation operates within the context of a `survey::svydesign`/`svrepdesign` object. Steps are recorded lazily, preserving sampling design metadata (weights, strata, PSUs, replicate weights) through every mutation, and materialized together via `bake_steps()`. This is fundamentally different from applying dplyr or data.table operations to survey data, where the design context is lost and must be manually re-attached.

  **Workflow automation**: The Recipe system bundles transformation pipelines into portable, versioned objects with metadata (survey type, edition, variable dependencies). Recipes can be published to a registry (local JSON or MongoDB-backed REST API), discovered by other researchers, validated against a survey's schema before execution, and applied to new editions of the same survey. A STATA `.do` file transpiler converts legacy processing scripts into Recipe objects, bridging the gap between existing institutional workflows and reproducible R pipelines. The combination — domain-aware steps, portable recipes, a discovery registry, and a legacy transpiler — is what makes metasurvey unique: each piece exists in isolation elsewhere, but no package integrates them into a single end-to-end infrastructure for survey data processing.

  Government statistical offices (INEs), academic researchers in economics, demography, and social sciences working with household survey microdata. The primary use case is reproducible processing of recurring national surveys — for example, Uruguay's ECH, Argentina's EPH, Brazil's PNADc, Chile's CASEN, or the US CPS.

  A critical distinction is that metasurvey is **not** a package for a single survey. Packages like `ech`, `eph`, `PNADcIBGE`, or `lodown` are tied to one specific survey and hardcode its variable names, editions, and conventions. metasurvey is survey-agnostic infrastructure: it provides the abstraction layer (steps, recipes, workflows, provenance) that works for *any* complex survey. What makes it useful for a specific survey is a Recipe contributed by the community — not code baked into the package. This means metasurvey scales through community contributions: as researchers publish recipes for their surveys, the ecosystem grows without requiring changes to the package itself.

  Today, survey processing at statistical offices and universities across Latin America is done with ad-hoc, non-reproducible scripts (often in STATA). metasurvey provides the infrastructure to standardize, share, and reproduce these pipelines across survey editions and across teams.

- Are there other R packages that accomplish the same thing? If so, how does yours differ?

  No single package covers the same ground. metasurvey's value is in the integration of capabilities that individually overlap with existing tools:

  - **`survey`** (Lumley): The estimation engine. metasurvey wraps it and adds everything *upstream* — the data processing pipeline that prepares microdata before `svymean()`/`svytotal()` can run. `survey` assumes clean, analysis-ready data; metasurvey handles the messy path from raw microdata to that point.
  - **`srvyr`**: Tidyverse syntax for `survey` objects. Focuses on the analysis phase (after data is clean). metasurvey focuses on the munging phase (before analysis) and is complementary.
  - **`recipes`** (tidymodels): General-purpose preprocessing pipelines for modeling. metasurvey's Recipe system is domain-specific: recipes carry survey metadata (type, edition, variable dependencies), can be published/discovered in a registry, and validated against a survey's schema. The step semantics are also different — they operate on design-aware survey objects, not bare data frames.
  - **`ech`/`eph`/`PNADcIBGE`/`lodown`**: Country- or survey-specific packages that hardcode variable names, download URLs, and processing logic for one survey. metasurvey is the *infrastructure layer* underneath: survey-agnostic steps, recipes, and workflows that any of these packages could use (or that researchers can use directly without needing a survey-specific package at all).
  - **`drake`/`targets`**: General workflow orchestration. metasurvey's workflow system is narrower and survey-specific: it wraps `svymean`/`svytotal`/`svyratio`/`svyby` calls with provenance tracking and ties results back to the recipe that produced the input data.

  The combination of design-aware lazy steps + portable recipes with a discovery registry + a STATA transpiler + provenance tracking is unique. Each component has analogues, but no package assembles them into an integrated, survey-agnostic processing infrastructure that scales through community-contributed recipes.

- Any other questions or issues we should be aware of?

  The package includes optional components (REST API client, MongoDB backend, Shiny explorer app) that extend the core for collaborative use. The core step pipeline and recipe system work entirely locally with no external dependencies beyond `survey` and `data.table`. We chose R6 classes because survey objects are inherently stateful (accumulating steps, holding mutable design objects) — we are happy to discuss this design choice.

- [x] I have read the [rOpenSci statistical software guide](https://stats-devguide.ropensci.org/) and confirm this package is **not** a statistical software package.
- [x] I have run `pkgcheck::pkgcheck()` locally and addressed all issues.
- [x] I have run `R CMD check` with no errors, warnings, or notes.
