### 2026-02-17 5fd7022
ci: add pkgcheck GitHub Action for rOpenSci submission
