+++ articles/workflows-and-estimation-es.md
@@ -395,7 +395,7 @@ permite rastrear cualquier estimacion hasta los datos originales.
 prov <- provenance(svy_full)
 prov
 #> ── Data Provenance ─────────────────────────────────────────────────────────────
-#> Loaded: 2026-02-17T16:06:29 
+#> Loaded: 2026-02-17T16:49:03 
 #> Initial rows: 200 
 #> 
 #> Environment:
