### 2026-02-17 650f609
refactor: move httr2 to Suggests, rewrite DESCRIPTION
- Move httr2 from Imports to Suggests with requireNamespace guards
- Add .check_httr2() helper for ANDA functions
- Rewrite Description field to avoid "Wraps" (rOpenSci policy)
- Add lifecycle experimental badges to anda_download_microdata, anda_variables
