### 2026-02-17 3e44013
docs: add lifecycle badges, Related Work, reduce vignettes
- Add experimental lifecycle badges to transpile_stata, transpile_stata_module, transpile_coverage
- Add Related Work comparison table to README
- Exclude 4 non-core vignettes from CRAN bundle (pkgdown-only articles)
- Minor roxygen formatting in publish_recipe
