+++ llms.txt
@@ -316,6 +316,23 @@ for the full reference.
 
 ------------------------------------------------------------------------
 
+## Related work
+
+| Package                                               | Focus                             | metasurvey adds                                        |
+|-------------------------------------------------------|-----------------------------------|--------------------------------------------------------|
+| [survey](https://cran.r-project.org/package=survey)   | Sampling designs and estimation   | Lazy step pipeline, recipe system, rotating panels     |
+| [srvyr](https://cran.r-project.org/package=srvyr)     | dplyr-style interface to survey   | Portable recipes, workflow registry, panel support     |
+| [recipes](https://cran.r-project.org/package=recipes) | Feature engineering for modelling | Survey-aware steps, complex designs, community sharing |
+| [eph](https://cran.r-project.org/package=eph)         | Argentina’s EPH survey            | Survey-agnostic: works with any household survey       |
+| [targets](https://cran.r-project.org/package=targets) | General pipeline orchestration    | Domain-specific steps, built-in survey semantics       |
+
+metasurvey is **not** a wrapper around `survey`. It adds a
+reproducibility layer (steps, recipes, workflows) that is
+survey-agnostic: the same pipeline processes ECH, EPH, CASEN, PNAD-C,
+CPS, ENIGH, or DHS data without survey-specific code.
+
+------------------------------------------------------------------------
+
 ## Citation
 
 To cite metasurvey in publications use:
@@ -640,20 +657,21 @@ REST client for the recipe registry
 INE Uruguay ANDA catalog metadata
 
 - [`anda_variables()`](https://metasurveyr.github.io/metasurvey/reference/anda_variables.md)
-  : Query ANDA variable metadata from the API
+  **\[experimental\]** : Query ANDA variable metadata from the API
 - [`anda_download_microdata()`](https://metasurveyr.github.io/metasurvey/reference/anda_download_microdata.md)
-  : Download ECH microdata from ANDA5
+  **\[experimental\]** : Download ECH microdata from ANDA5
 
 ## STATA Transpiler
 
 Convert STATA do-files to metasurvey recipes
 
 - [`transpile_stata()`](https://metasurveyr.github.io/metasurvey/reference/transpile_stata.md)
-  : Transpile a STATA .do file to metasurvey steps
+  **\[experimental\]** : Transpile a STATA .do file to metasurvey steps
 - [`transpile_stata_module()`](https://metasurveyr.github.io/metasurvey/reference/transpile_stata_module.md)
-  : Transpile and group do-files by thematic module
+  **\[experimental\]** : Transpile and group do-files by thematic module
 - [`transpile_coverage()`](https://metasurveyr.github.io/metasurvey/reference/transpile_coverage.md)
-  : Analyze transpilation coverage for STATA do-files
+  **\[experimental\]** : Analyze transpilation coverage for STATA
+  do-files
 - [`parse_do_file()`](https://metasurveyr.github.io/metasurvey/reference/parse_do_file.md)
   : Parse a STATA .do file into structured commands
 - [`parse_stata_labels()`](https://metasurveyr.github.io/metasurvey/reference/parse_stata_labels.md)
