### 2026-02-17 29970b3
feat: add convey package support for inequality/poverty measures
- Add cat_estimation.cvystat() handler for convey return objects
- Update cat_estimation() dispatcher to recognize cvystat class
- Add convey to Suggests in DESCRIPTION
- Add pkgdown-only vignette with Gini, Atkinson, FGT examples
- Add 4 tests for cvystat extraction and convey integration
