+++ vignettes/convey-inequality.Rmd
@@ -1,0 +1,287 @@
+---
+title: "Poverty and Inequality with convey"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Poverty and Inequality with convey}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  warning = FALSE,
+  message = FALSE
+)
+can_run <- requireNamespace("convey", quietly = TRUE)
+```
+
+## Introduction
+
+The [convey](https://www.convey-r.org/) package by Guilherme Jacob, Anthony
+Damico, and Djalma Pessoa implements poverty and inequality indicators for
+complex survey data. It works with `survey::svydesign` objects --- the same
+objects that metasurvey wraps inside `Survey` objects.
+
+This vignette shows how to use `convey` functions inside `workflow()` to
+compute Gini coefficients, at-risk-of-poverty rates, FGT indices, and other
+distributional measures, all with proper standard errors and CVs.
+
+For the full reference on every measure, see the
+[convey book](https://www.convey-r.org/).
+
+## Setup
+
+We use the `api` dataset from the `survey` package. The `api00` variable
+(Academic Performance Index score in 2000) serves as our continuous variable
+for inequality measures, and `meals` (percent of students eligible for
+subsidized meals) works as an income-like proxy.
+
+```{r setup, eval = can_run}
+library(metasurvey)
+library(survey)
+library(convey)
+library(data.table)
+
+data(api, package = "survey")
+dt <- data.table(apistrat)
+
+svy <- Survey$new(
+  data    = dt,
+  edition = "2000",
+  type    = "api",
+  psu     = NULL,
+  engine  = "data.table",
+  weight  = add_weight(annual = "pw")
+)
+```
+
+### Preparing the design for convey
+
+Before using any `convey` function, the underlying design must be prepared
+with `convey_prep()`. Build the design with `ensure_design()` and then
+replace the estimation-type entry:
+
+```{r convey-prep, eval = can_run}
+svy$ensure_design()
+svy$design[["annual"]] <- convey_prep(svy$design[["annual"]])
+```
+
+## Inequality Measures
+
+### Gini coefficient
+
+The Gini index measures overall inequality on a 0--1 scale:
+
+```{r gini, eval = can_run}
+gini <- workflow(
+  list(svy),
+  convey::svygini(~api00, na.rm = TRUE),
+  estimation_type = "annual"
+)
+
+gini
+```
+
+### Atkinson index
+
+The Atkinson index uses an inequality aversion parameter `epsilon`.
+Higher epsilon gives more weight to the lower tail:
+
+```{r atkinson, eval = can_run}
+atk_05 <- workflow(
+  list(svy),
+  convey::svyatk(~api00, epsilon = 0.5),
+  estimation_type = "annual"
+)
+
+atk_1 <- workflow(
+  list(svy),
+  convey::svyatk(~api00, epsilon = 1),
+  estimation_type = "annual"
+)
+
+rbind(atk_05, atk_1)
+```
+
+### Quintile share ratio (QSR)
+
+The QSR compares income at the top 20% with the bottom 20%:
+
+```{r qsr, eval = can_run}
+qsr <- workflow(
+  list(svy),
+  convey::svyqsr(~api00, na.rm = TRUE),
+  estimation_type = "annual"
+)
+
+qsr
+```
+
[…]
