+++ _pkgdown.yml
@@ -220,6 +220,7 @@ articles:
   - ech-demographics-recipe
   - ech-case-study
   - international-surveys
+  - convey-inequality
 - title: "Español"
   desc: "Documentación en español"
   contents:
