### 2026-02-17 5b32e35
docs: add convey support to NEWS.md
