+++ reference/transpile_coverage.md
@@ -42,8 +42,8 @@ tf <- tempfile(fileext = ".do")
 writeLines(c("gen x = 1", "replace x = 2 if y == 3", "drop z"), tf)
 transpile_coverage(tf)
 #>                  file                                path total_commands
-#> 1 file1e2c3ae72c68.do /tmp/RtmpD3FlmY/file1e2c3ae72c68.do              2
-#> 2               TOTAL /tmp/RtmpD3FlmY/file1e2c3ae72c68.do              2
+#> 1 file1e7414881b47.do /tmp/RtmphMunPh/file1e7414881b47.do              2
+#> 2               TOTAL /tmp/RtmphMunPh/file1e7414881b47.do              2
 #>   translated skipped manual_review coverage_pct
 #> 1          2       0             0          100
 #> 2          2       0             0          100
