+++ reference/read_recipe.md
@@ -46,7 +46,7 @@ r <- recipe(
 )
 f <- tempfile(fileext = ".json")
 save_recipe(r, f)
-#> The recipe has been saved in /tmp/RtmphMunPh/file1e74351038b.json
+#> The recipe has been saved in /tmp/RtmpwU7woC/file1e4a1632863f.json
 r2 <- read_recipe(f)
 #> Warning: Failed to parse recipe steps: invalid length 0 argument. Using raw strings as fallback.
 r2
