+++ articles/workflows-and-estimation-es.md
@@ -395,7 +395,7 @@ permite rastrear cualquier estimacion hasta los datos originales.
 prov <- provenance(svy_full)
 prov
 #> ── Data Provenance ─────────────────────────────────────────────────────────────
-#> Loaded: 2026-02-17T22:22:48 
+#> Loaded: 2026-02-17T22:27:14 
 #> Initial rows: 200 
 #> 
 #> Environment:
