+++ articles/ech-demographics-recipe.md
@@ -353,13 +353,13 @@ cat(readLines(f, n = 15), sep = "\n")
 #>   "description": "Harmonized demographics: sex, age group, relationship, education level, and geographic area.",
 #>   "topic": "demographics",
 #>   "doi": {},
-#>   "id": "r_1771366901_663",
+#>   "id": "r_1771367168_663",
 #>   "version": "1.0.0",
 #>   "downloads": 0,
 #>   "categories": [],
 #>   "certification": {
 #>     "level": "community",
-#>     "certified_at": "2026-02-17 22:21:41.946556"
+#>     "certified_at": "2026-02-17 22:26:08.404974"
 ```
 
 ## Applying to a new edition
