+++ reference/save_workflow.md
@@ -45,5 +45,5 @@ wf <- RecipeWorkflow$new(
 )
 f <- tempfile(fileext = ".json")
 save_workflow(wf, f)
-#> Workflow saved to /tmp/RtmpbQUJx4/file1e2458ffc694.json
+#> Workflow saved to /tmp/RtmpctN7P6/file1e03a68ccbd.json
 ```
