+++ reference/transpile_coverage.md
@@ -42,8 +42,8 @@ tf <- tempfile(fileext = ".do")
 writeLines(c("gen x = 1", "replace x = 2 if y == 3", "drop z"), tf)
 transpile_coverage(tf)
 #>                  file                                path total_commands
-#> 1 file1e0367887de1.do /tmp/RtmpctN7P6/file1e0367887de1.do              2
-#> 2               TOTAL /tmp/RtmpctN7P6/file1e0367887de1.do              2
+#> 1 file1e497c089f46.do /tmp/Rtmp5MQdxI/file1e497c089f46.do              2
+#> 2               TOTAL /tmp/Rtmp5MQdxI/file1e497c089f46.do              2
 #>   translated skipped manual_review coverage_pct
 #> 1          2       0             0          100
 #> 2          2       0             0          100
