### 2026-02-18 30e74c3
fix: resolve invalid file URIs flagged by CRAN
- Replace relative file URIs in README.md and CONTRIBUTING.md with full GitHub URLs
- Wrap bare URL in roxygen @description with \url{} in R/utils.R
- Fill empty @description fields in R/api_client.R
- Add Rplots.pdf and convey-inequality_files/ to .Rbuildignore
