+++ man/load_survey_example.Rd
@@ -31,7 +31,7 @@ locally in a temporary file to avoid repeated downloads
 in the same session.
 
 Available survey types and editions can be found at:
-https://github.com/metasurveyr/metasurvey_data
+\url{https://github.com/metasurveyr/metasurvey_data}
 }
 \examples{
 \dontrun{
