+++ CONTRIBUTING.md
@@ -52,5 +52,5 @@ their prioritization.
 ## Code of Conduct
 
 Please note that this project is released with a
-[Contributor Code of Conduct](CODE_OF_CONDUCT.md). By participating in this
+[Contributor Code of Conduct](https://github.com/metasurveyr/metasurvey/blob/main/CODE_OF_CONDUCT.md). By participating in this
 project you agree to abide by its terms.
