### 2026-02-18 de710d8
docs: print provenance object in convey vignette
