### 2026-02-18 4075c93
docs: update cran-comments for resubmission
