+++ reference/get_backend.md
@@ -41,6 +41,6 @@ backend
 #>     search: function (query) 
 #>     type: local
 #>   Private:
-#>     .path: /tmp/Rtmp5MQdxI/file1e497ada3f18.json
+#>     .path: /tmp/RtmptjUtmc/file1e715c2a607a.json
 #>     .registry: RecipeRegistry, R6
 ```
