+++ reference/api_get_workflow.md
@@ -1,6 +1,6 @@
 # Get a single workflow by ID
 
-Get a single workflow by ID
+Retrieve a workflow from the API by its ID.
 
 ## Usage
 
