+++ CONTRIBUTING.md
@@ -54,5 +54,5 @@ the team manages their prioritization.
 ## Code of Conduct
 
 Please note that this project is released with a [Contributor Code of
-Conduct](https://metasurveyr.github.io/metasurvey/CODE_OF_CONDUCT.md).
+Conduct](https://github.com/metasurveyr/metasurvey/blob/main/CODE_OF_CONDUCT.md).
 By participating in this project you agree to abide by its terms.
