### 2026-02-18 a54cc35
docs: note INE Uruguay URL timeout in cran-comments
