### 2026-02-18 ea9437b
docs: replace broken DOI with FCEA ECH page
