+++ articles/ech-demographics-recipe-es.md
@@ -355,13 +355,13 @@ cat(readLines(f, n = 15), sep = "\n")
 #>   "description": "Demografia armonizada: sexo, grupo de edad, relacion, nivel educativo y area geografica.",
 #>   "topic": "demographics",
 #>   "doi": {},
-#>   "id": "r_1771375233_663",
+#>   "id": "r_1771375858_663",
 #>   "version": "1.0.0",
 #>   "downloads": 0,
 #>   "categories": [],
 #>   "certification": {
 #>     "level": "community",
-#>     "certified_at": "2026-02-18 00:40:33.122004"
+#>     "certified_at": "2026-02-18 00:50:58.091818"
 ```
 
 ## Aplicando a una nueva edicion
