+++ reference/get_backend.md
@@ -41,6 +41,6 @@ backend
 #>     search: function (query) 
 #>     type: local
 #>   Private:
-#>     .path: /tmp/Rtmp4K7R7L/file1e27160b1ef2.json
+#>     .path: /tmp/RtmpRWKZLg/file20666464b658.json
 #>     .registry: RecipeRegistry, R6
 ```
