+++ _pkgdown.yml
@@ -14,7 +14,7 @@ navbar:
 
 authors:
   Mauro Loprete:
-    href: https://www.mauroloprete.com
+    href: https://iesta.fcea.udelar.edu.uy/integrantes/loprete-mauro/
   Natalia da Silva:
     href: https://natydasilva.github.io/natydasilva/
   Fabricio Machado:
