+++ reference/print.metasurvey_provenance.md
@@ -41,10 +41,10 @@ s <- step_compute(s, age2 = age * 2)
 s <- bake_steps(s)
 print(provenance(s))
 #> ── Data Provenance ─────────────────────────────────────────────────────────────
-#> Loaded: 2026-02-18T00:54:48 
+#> Loaded: 2026-02-18T01:05:21 
 #> 
 #> Pipeline:
-#>   1. step_1 Compute: age2  N=48 [0.0ms]
+#>   1. step_1 Compute: age2  N=48 [1.0ms]
 #> 
 #> Environment:
 #>   metasurvey: 0.0.21 
