### 2026-02-19 #752 @emilyriederer
Hi @mauroloprete ! Thank you for your submission. I think this definitely seems like a fit for rOpenSci.

Two questions for you:

- would you be open to removing the "data munging" scope? I think the actual wrangling capability of the package are narrower in scope (i.e. the `step_`) functions versus the overall structure for a survey workflow. So, the fit for the latter seems far stronger.
- I do note this is quite a large package at the moment with so many types of functionality! I think we will need to give reviewers a bit more structure to navigate the repo. Would you be open to adding `README`s to your `R/` and `inst/` directories to help readers better understand what they contain and how to navigate them?
