### 2026-02-19 6742232
docs: add README guides for R/ and inst/ directories
- R/README.md: architecture diagram, problem statement, why R6, and
  per-layer descriptions with domain context
- inst/README.md: tree diagram grouping contents by purpose (sample
  data, deployable services, transpiler fixtures, registry bootstrap)
- Remove empty inst/shiny-auth.R
