+++ R/README.md
@@ -1,0 +1,132 @@
+# `R/` Directory Guide
+
+The source code is organized in six layers that mirror the package's architecture:
+
+```
+                      STATA .do files        ANDA catalog
+                           |                     |
+                       transpile()          fetch metadata
+                           |                     |
+                           v                     v
+                  +-----------------+   +-----------------+
+                  |   5. Interop    |   |   5. Interop    |
+                  |  stata_parser   |   |    anda.R       |
+                  |  stata_mappings |   |                 |
+                  +--------+--------+   +--------+--------+
+                           |                     |
+             produces Recipe|          enriches Survey
+                           v                     v
++-----------------------------------------------------------+
+|              4. Registry & Backends                       |
+|       publish, discover, share (JSON / MongoDB)           |
+|  RecipeRegistry · WorkflowRegistry · api_client           |
++----------------------------+------------------------------+
+                             |
+              stores & retrieves
+                             |
++----------------------------v------------------------------+
+|            3. Recipes & Workflows                         |
+|  Recipe = portable bundle of steps + metadata             |
+|  Workflow = estimation (svymean/svytotal) + provenance    |
++----------------------------+------------------------------+
+                             |
+                 bundles & executes
+                             |
++----------------------------v------------------------------+
+|              2. Step Pipeline                             |
+|  step_compute · step_recode · step_rename                 |
+|  step_remove  · step_join   · step_filter                 |
+|                    bake_steps()                            |
++----------------------------+------------------------------+
+                             |
+                 operates on
+                             |
++----------------------------v------------------------------+
+|              1. Survey Objects                            |
+|  Survey · RotativePanelSurvey · PoolSurvey               |
+|         data.table + survey::svydesign                    |
++-----------------------------------------------------------+
+```
+
+## The problem metasurvey solves
+
+Research institutes and university departments that work with national household surveys (ECH in Uruguay, EPH in Argentina, CASEN in Chile, CPS in the US) typically follow a well-defined methodology to process raw microdata: harmonize variables across editions, construct derived indicators (income, poverty lines, employment categories), apply sampling weights, and produce estimates. This processing is implemented in STATA `.do` files that accumulate years of institutional knowledge.
+
+The results — processed datasets, indicator tables — are shared informally, often by email, as flat files or spreadsheets. External researchers, policy monitors, and government reports then use these processed variables to make decisions, publish papers, and build dashboards. But the processing pipeline itself is opaque: at best, you receive a `.do` file that rarely runs outside the original machine (hardcoded paths, missing dependencies, undocumented STATA version requirements). There is no standard way to inspect what transformations were applied, validate them against a new survey edition, or reproduce them independently.
+
+metasurvey replaces this workflow with transparent, reproducible infrastructure: every transformation is recorded as a step, bundled into a portable recipe with metadata and dependency tracking, and published to a shared registry where others can discover, validate, and apply it — without needing STATA or chasing email attachments.
+
+## Why R6?
+
[…]
