### 2026-02-19 94f18d3
fix: add eval=FALSE to vignettes needing external data
- ech-demographics-recipe.Rmd and stata-transpiler.Rmd: set eval=FALSE
  so R CMD check does not fail when example data is unavailable
- Remove stale comments in shiny app
