### 2026-02-19 510139a
docs: add navbar dropdowns for articles and reference
- Get Started as explicit navbar link
- Articles dropdown grouped by topic (Core, Advanced, Infrastructure,
  Interoperability, Case Studies) with Spanish section
- Reference dropdown with links to each topic section
