### 2026-02-19 52e767a
Update README with database script details
