### 2026-02-20 d12e29f
fix: remove per-chunk eval overrides in international-surveys
Per-chunk eval = requireNamespace(...) overrode the global eval = FALSE,
causing CI failures when eph was installed but data.table was not loaded.
The global eval = FALSE now controls all chunks.
