+++ vignettes/international-surveys-es.Rmd
@@ -118,7 +118,7 @@ laboral de Argentina. El paquete `eph`
 ([CRAN](https://cran.r-project.org/package=eph),
 [GitHub](https://github.com/ropensci/eph)) incluye una base de juguete.
 
-```{r eph, eval = requireNamespace("eph", quietly = TRUE)}
+```{r eph}
 library(eph)
 
 data("toybase_individual_2016_04", package = "eph")
@@ -228,7 +228,7 @@ por conglomerados. El paquete `PNADcIBGE`
 ([CRAN](https://cran.r-project.org/package=PNADcIBGE)) incluye microdatos
 de ejemplo.
 
-```{r pnadc, eval = requireNamespace("PNADcIBGE", quietly = TRUE)}
+```{r pnadc}
 library(PNADcIBGE)
 
 dt_pnadc <- data.table(read_pnadc(
@@ -277,7 +277,7 @@ laboral de Estados Unidos. El paquete `ipumsr`
 ([CRAN](https://cran.r-project.org/package=ipumsr),
 [GitHub](https://github.com/ipums/ipumsr)) incluye un extracto de CPS.
 
-```{r cps, eval = requireNamespace("ipumsr", quietly = TRUE)}
+```{r cps}
 library(ipumsr)
 
 ddi <- read_ipums_ddi(
@@ -379,7 +379,7 @@ El paquete `rdhs` ([CRAN](https://cran.r-project.org/package=rdhs),
 Los datasets modelo (sin necesidad de autenticacion) estan disponibles en
 <https://dhsprogram.com/data/model-datasets.cfm>.
 
-```{r dhs, eval = requireNamespace("haven", quietly = TRUE)}
+```{r dhs}
 library(haven)
 
 # Descargar el modelo Individual Recode (sin credenciales)
