+++ articles/international-surveys.md
@@ -83,12 +83,6 @@ workflow(
   survey::svymean(~HT11, na.rm = TRUE),
   estimation_type = "annual"
 )
-#>                     stat    value       se         cv confint_lower
-#>                   <char>    <num>    <num>      <num>         <num>
-#> 1: survey::svymean: HT11 107869.1 3473.836 0.03220417      101060.5
-#>    confint_upper
-#>            <num>
-#> 1:      114677.7
 ```
 
 For the complete ECH pipeline, see
@@ -142,12 +136,6 @@ workflow(
   survey::svymean(~employed, na.rm = TRUE),
   estimation_type = "quarterly"
 )
-#>                         stat     value         se         cv confint_lower
-#>                       <char>     <num>      <num>      <num>         <num>
-#> 1: survey::svymean: employed 0.4378169 0.01668347 0.03810603     0.4051179
-#>    confint_upper
-#>            <num>
-#> 1:     0.4705159
 ```
 
 To download actual EPH microdata use
@@ -252,12 +240,6 @@ workflow(
   survey::svymean(~age, na.rm = TRUE),
   estimation_type = "quarterly"
 )
-#>                    stat    value       se         cv confint_lower
-#>                  <char>    <num>    <num>      <num>         <num>
-#> 1: survey::svymean: age 35.55343 0.856023 0.02407708      33.87566
-#>    confint_upper
-#>            <num>
-#> 1:      37.23121
 ```
 
 For real PNADc microdata:
@@ -310,12 +292,6 @@ workflow(
   survey::svymean(~INCTOT, na.rm = TRUE),
   estimation_type = "annual"
 )
-#>                       stat     value      se         cv confint_lower
-#>                     <char>     <num>   <num>      <num>         <num>
-#> 1: survey::svymean: INCTOT 191645236 4277528 0.02232004     183261435
-#>    confint_upper
-#>            <num>
-#> 1:     200029038
 ```
 
 IPUMS data requires a free account at <https://cps.ipums.org>. The DDI
@@ -370,12 +346,6 @@ workflow(
   survey::svymean(~income_pc, na.rm = TRUE),
   estimation_type = "annual"
 )
-#>                          stat    value       se        cv confint_lower
-#>                        <char>    <num>    <num>     <num>         <num>
-#> 1: survey::svymean: income_pc 11928.34 1473.022 0.1234893      9041.267
-#>    confint_upper
-#>            <num>
-#> 1:      14815.41
 ```
 
 ## DHS – International
@@ -484,9 +454,7 @@ my_recipe <- steps_to_recipe(
 
 doc <- my_recipe$doc()
 cat("Inputs:", paste(doc$input_variables, collapse = ", "), "\n")
-#> Inputs: age
 cat("Outputs:", paste(doc$output_variables, collapse = ", "), "\n")
-#> Outputs: indicator, age_group
 ```
 
 Recipes capture *what* transformations to apply, not *which* survey they
