### 2026-02-20 7811574
fix: restore eval in international-surveys vignettes
- Remove global eval=FALSE, chunks now run by default
- eph/pnadc/cps chunks: eval conditional on package availability
- dhs chunk: eval=FALSE (requires external download from dhsprogram.com)
- ech/enigh/recipe-portability: always eval (bundled or synthetic data)
