+++ vignettes/international-surveys.Rmd
@@ -11,7 +11,6 @@ vignette: >
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
-  eval = FALSE,
   warning = FALSE,
   message = FALSE
 )
@@ -116,7 +115,7 @@ force survey. The `eph` package
 ([CRAN](https://cran.r-project.org/package=eph),
 [GitHub](https://github.com/ropensci/eph)) includes a bundled toybase.
 
-```{r eph}
+```{r eph, eval = has_eph}
 library(eph)
 
 data("toybase_individual_2016_04", package = "eph")
@@ -225,7 +224,7 @@ The `PNADcIBGE` package
 ([CRAN](https://cran.r-project.org/package=PNADcIBGE)) includes bundled
 example microdata.
 
-```{r pnadc}
+```{r pnadc, eval = has_pnad}
 library(PNADcIBGE)
 
 dt_pnadc <- data.table(read_pnadc(
@@ -272,7 +271,7 @@ The CPS (Current Population Survey) is the US monthly labor force survey.
 The `ipumsr` package ([CRAN](https://cran.r-project.org/package=ipumsr),
 [GitHub](https://github.com/ipums/ipumsr)) includes a bundled CPS extract.
 
-```{r cps}
+```{r cps, eval = has_ipumsr}
 library(ipumsr)
 
 ddi <- read_ipums_ddi(
@@ -373,7 +372,7 @@ The `rdhs` package ([CRAN](https://cran.r-project.org/package=rdhs),
 datasets (no authentication needed) are available at
 <https://dhsprogram.com/data/model-datasets.cfm>.
 
-```{r dhs}
+```{r dhs, eval = FALSE}
 library(haven)
 
 # Download the model Individual Recode (no credentials needed)
