+++ articles/international-surveys-es.md
@@ -85,6 +85,12 @@ workflow(
   survey::svymean(~HT11, na.rm = TRUE),
   estimation_type = "annual"
 )
+#>                     stat    value       se         cv confint_lower
+#>                   <char>    <num>    <num>      <num>         <num>
+#> 1: survey::svymean: HT11 107869.1 3473.836 0.03220417      101060.5
+#>    confint_upper
+#>            <num>
+#> 1:      114677.7
 ```
 
 Para el pipeline completo de la ECH, ver
@@ -138,6 +144,12 @@ workflow(
   survey::svymean(~employed, na.rm = TRUE),
   estimation_type = "quarterly"
 )
+#>                         stat     value         se         cv confint_lower
+#>                       <char>     <num>      <num>      <num>         <num>
+#> 1: survey::svymean: employed 0.4378169 0.01668347 0.03810603     0.4051179
+#>    confint_upper
+#>            <num>
+#> 1:     0.4705159
 ```
 
 Para descargar microdatos reales de la EPH:
@@ -243,6 +255,12 @@ workflow(
   survey::svymean(~age, na.rm = TRUE),
   estimation_type = "quarterly"
 )
+#>                    stat    value       se         cv confint_lower
+#>                  <char>    <num>    <num>      <num>         <num>
+#> 1: survey::svymean: age 35.55343 0.856023 0.02407708      33.87566
+#>    confint_upper
+#>            <num>
+#> 1:      37.23121
 ```
 
 Para microdatos reales de la PNADc:
@@ -294,6 +312,12 @@ workflow(
   survey::svymean(~INCTOT, na.rm = TRUE),
   estimation_type = "annual"
 )
+#>                       stat     value      se         cv confint_lower
+#>                     <char>     <num>   <num>      <num>         <num>
+#> 1: survey::svymean: INCTOT 191645236 4277528 0.02232004     183261435
+#>    confint_upper
+#>            <num>
+#> 1:     200029038
 ```
 
 Los datos de IPUMS requieren una cuenta gratuita en
@@ -349,6 +373,12 @@ workflow(
   survey::svymean(~income_pc, na.rm = TRUE),
   estimation_type = "annual"
 )
+#>                          stat    value       se        cv confint_lower
+#>                        <char>    <num>    <num>     <num>         <num>
+#> 1: survey::svymean: income_pc 11928.34 1473.022 0.1234893      9041.267
+#>    confint_upper
+#>            <num>
+#> 1:      14815.41
 ```
 
 ## DHS – Internacional
@@ -458,7 +488,9 @@ my_recipe <- steps_to_recipe(
 
 doc <- my_recipe$doc()
 cat("Entradas:", paste(doc$input_variables, collapse = ", "), "\n")
+#> Entradas: age
 cat("Salidas:", paste(doc$output_variables, collapse = ", "), "\n")
+#> Salidas: indicator, age_group
 ```
 
 Los recipes capturan *que* transformaciones aplicar, no *de que*
