+++ reference/provenance_to_json.md
@@ -44,7 +44,7 @@ provenance_to_json(prov)
 #> {
 #>   "source": {
 #>     "path": null,
-#>     "timestamp": "2026-02-20T02:06:16",
+#>     "timestamp": "2026-02-20T02:17:56",
 #>     "initial_n": 5,
 #>     "hash": null
 #>   },
