### 2026-02-20 c96c92c
fix: remove examples from unexported internal functions
CRAN rejected: "examples for unexported function cat_recipes()".
Removed @examples from all 6 internal classes/functions:
- cat_recipes() in R/survey.R
- Step class in R/Step.R
- RecipeBackend class in R/RecipeAPI.R
- RecipeRegistry class in R/RecipeRegistry.R
- WorkflowBackend class in R/WorkflowBackend.R
- WorkflowRegistry class in R/WorkflowRegistry.R
