+++ man/RecipeRegistry.Rd
@@ -7,13 +7,6 @@
 Local JSON-backed catalog for recipe
 discovery, ranking, and filtering.
 }
-\examples{
-# Use the tidy API instead:
-# set_backend("local", path = "recipes.json")
-# publish_recipe(my_recipe)
-# search_recipes("labor")
-
-}
 \keyword{internal}
 \section{Methods}{
 \subsection{Public methods}{
