+++ man/WorkflowRegistry.Rd
@@ -7,13 +7,6 @@
 Local JSON-backed catalog for workflow
 discovery, ranking, and filtering.
 }
-\examples{
-# Use the tidy API instead:
-# set_workflow_backend("local", path = "workflows.json")
-# publish_workflow(my_workflow)
-# search_workflows("labor")
-
-}
 \keyword{internal}
 \section{Methods}{
 \subsection{Public methods}{
