### 2026-02-20 c16f641
test: add guard against broken vignette cross-references
Prevents CRAN rejection by detecting relative HTML links in
included vignettes that point to excluded vignettes. Runs as
part of devtools::test() before submission.
