### 2026-02-20 9bdf494
docs: update cran-comments for resubmission
- Document fixes for invalid file URI and unexported examples
- Report 1 NOTE (new submission + INE Uruguay URL timeout)
- Add R/README.md to .Rbuildignore
