### 2026-02-21 #752 @emilyriederer
Thanks, @mauroloprete ! I appreciate the updates. I think those will be helpful.

We are also currently in the process of defining our approach to reviews for packages developed with the use of AI. Although the guidance is not formally published, we do plan to require authors to explicitly describe how, where, and why generative AI tools were used. (This is not a value judgement on using AI but instructive to think about how to approach the review).

I noticed numerous Claude artifacts in your repo. Could you please provide a high-level explanation to how you distinguished human-generated and AI-generated tasks while developing your package?
