### 2026-02-25 2493147
docs: add CRAN installation instructions to README
- Add install.packages("metasurvey") as primary installation method
- Move GitHub install to secondary "development version" option
- Update citation year to 2026 and link to CRAN page
