+++ articles/workflows-and-estimation-es.md
@@ -395,13 +395,13 @@ permite rastrear cualquier estimacion hasta los datos originales.
 prov <- provenance(svy_full)
 prov
 #> ── Data Provenance ─────────────────────────────────────────────────────────────
-#> Loaded: 2026-02-20T02:19:33 
+#> Loaded: 2026-02-25T12:03:10 
 #> Initial rows: 200 
 #> 
 #> Environment:
 #>   metasurvey: 0.0.21 
 #>   R: 4.5.2 
-#>   survey: 4.4.8
+#>   survey: 4.5
 ```
 
 El provenance tambien se adjunta a los resultados de
@@ -449,7 +449,7 @@ workflow_table(estimates)
 | Statistic                                  | Estimate | SE    | CI Lower | CI Upper | CV (%) | Quality   |
 | :svymean: api_growth                       | 32.89    | 2.158 | 28.66    | 37.12    | 6.6    | Very good |
 | :svymean: high_growth                      | 0.29     | 0.036 | 0.22     | 0.37     | 12.4   | Good      |
-| metasurvey 0.0.21 \| CI: 95% \| 2026-02-20 |          |       |          |          |        |           |
+| metasurvey 0.0.21 \| CI: 95% \| 2026-02-25 |          |       |          |          |        |           |
 
 Se puede personalizar la salida:
 
@@ -470,7 +470,7 @@ workflow_table(
 | Statistic                                  | Estimate | CI Lower | CI Upper | CV (%) | Quality   |
 | :svymean: api_growth                       | 32,89    | 28,66    | 37,12    | 6,6    | Very good |
 | :svymean: high_growth                      | 0,29     | 0,22     | 0,37     | 12,4   | Good      |
-| metasurvey 0.0.21 \| CI: 95% \| 2026-02-20 |          |          |          |        |           |
+| metasurvey 0.0.21 \| CI: 95% \| 2026-02-25 |          |          |          |        |           |
 
 Para estimaciones por dominio, la tabla detecta automaticamente las
 columnas de grupo:
@@ -485,7 +485,7 @@ workflow_table(by_school)
 | :svyby: api00                              | E     | 674.43   | 12.493 | 649.94   | 698.92   | 1.9    | Excellent |
 | :svyby: api00                              | H     | 625.82   | 15.341 | 595.75   | 655.89   | 2.5    | Excellent |
 | :svyby: api00                              | M     | 636.60   | 16.502 | 604.26   | 668.94   | 2.6    | Excellent |
-| metasurvey 0.0.21 \| CI: 95% \| 2026-02-20 |       |          |        |          |          |        |           |
+| metasurvey 0.0.21 \| CI: 95% \| 2026-02-25 |       |          |        |          |          |        |           |
 
 Se exporta a cualquier formato soportado por
 [`gt::gtsave()`](https://gt.rstudio.com/reference/gtsave.html):
