### 2026-02-28 2821616
docs: add recipe contribution guide and issue template
- Add "Contributing recipes" section to CONTRIBUTING.md with registry workflow
- Add "Contribute a Recipe" issue template for claiming surveys
- List target LATAM surveys: EPH, PNADc, CASEN, GEIH, ENIGH
