+++ CONTRIBUTING.md
@@ -26,6 +26,65 @@ The metasurvey team reviews discussions and decides which ideas get prioritized
 for implementation. Issues are reserved for confirmed bugs — the team manages
 their prioritization.
 
+## Contributing recipes
+
+The most impactful way to contribute to metasurvey is by **publishing a recipe**
+to the community registry. A recipe captures the full processing pipeline for a
+household survey — variable construction, recodes, filters — so other
+researchers can reproduce your work or apply it to new editions.
+
+### Publish a recipe to the registry
+
+```r
+library(metasurvey)
+
+# 1. Connect to the registry
+configure_api("https://metasurvey-api.onrender.com")
+
+# 2. Create an account (once)
+api_register("Your Name", "you@example.com", "your_password")
+
+# 3. Build your recipe
+svy <- survey_empty("eph", "2023T1")
+r <- recipe(
+  name = "Labor Market Indicators",
+  user = "you@example.com",
+  svy = svy,
+  description = "Activity, employment, and unemployment rates from EPH"
+)
+
+# 4. Publish
+api_publish_recipe(r)
+```
+
+### Browse existing recipes
+
+```r
+# From R
+configure_api("https://metasurvey-api.onrender.com")
+list_recipes()
+search_recipes("labor")
+
+# Or launch the Shiny explorer
+explore_recipes()
+```
+
+### Surveys we're looking for
+
+We are building a community collection for Latin American household surveys:
+
+| Survey | Country | Office |
+|--------|---------|--------|
+| EPH | Argentina | INDEC |
+| PNADc | Brazil | IBGE |
+| CASEN | Chile | MDS |
+| GEIH | Colombia | DANE |
+| ENIGH | Mexico | INEGI |
+
+If you work with any of these (or any other complex survey), publish your recipe
+and [open an issue](https://github.com/metasurveyr/metasurvey/issues/new?template=new_recipe.md)
+to let us know.
+
 ## Pull requests
 
 1. Fork the repo and create a branch from `develop`.
