+++ reference/save_recipe.md
@@ -50,5 +50,5 @@ r <- recipe(
 )
 f <- tempfile(fileext = ".json")
 save_recipe(r, f)
-#> The recipe has been saved in /tmp/RtmpkNIXgg/file26a235cc8d99.json
+#> The recipe has been saved in /tmp/Rtmpf7J2mh/file204911f82f41.json
 ```
