+++ reference/workflow_table.md
@@ -163,6 +163,6 @@ Quality
 
 Do not publish
 
-metasurvey 0.0.21 \| CI: 95% \| 2026-02-25
+metasurvey 0.0.21 \| CI: 95% \| 2026-02-28
 
 \# }

[…]
