### 2026-03-01 cc86cb6
chore: overhaul pre-commit hooks and add Makefile
- Add lintr, trailing-whitespace, check-yaml, detect-private-key hooks
- Add no-commit-to-branch protection for main
- Remove roxygenize hook (needs full renv, run make document instead)
- Add forbid-secrets hook to prevent .env commits
- Add Makefile with dev-setup, test, check, docker targets
- Add GitHub Actions workflow for pre-commit CI
- Add PR template
- Gitignore testthat-problems.rds artifacts
