+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,17 @@
+## Summary
+
+Brief description of what this PR does.
+
+Closes #
+
+## Changes
+
+-
+
+## Checklist
+
+- [ ] `devtools::document()` — roxygen docs regenerated
+- [ ] `devtools::test()` — all tests pass
+- [ ] `devtools::check()` — 0 errors, 0 warnings
+- [ ] New/changed exported functions have `@return`, `@examples`, `@param`
+- [ ] NEWS.md updated (if user-facing change)
