### 2026-03-01 29baec4
style: apply tidyverse formatting across R sources
Ran styler::style_pkg() — brace placement, trailing commas,
line breaks in long function calls.
