### 2026-03-01 af535c2
docs: rewrite CONTRIBUTING with dev environment guide
- Add install instructions for macOS, Ubuntu, Windows (WSL2)
- Add quick start section with make targets
- Document pre-commit hooks table
- Add architecture overview (R6 classes, patterns)
- Document Docker services and seed data
- Mention no-rlang rule and print() in tests
- Fix trailing newline in inst/README.md
- Normalize line endings in inst/scripts/README.md
