+++ CONTRIBUTING.md
@@ -85,13 +85,184 @@ If you work with any of these (or any other complex survey), publish your recipe
 and [open an issue](https://github.com/metasurveyr/metasurvey/issues/new?template=new_recipe.md)
 to let us know.
 
+## Development environment
+
+### Installing the stack
+
+<details>
+<summary><strong>macOS</strong></summary>
+
+```bash
+# 1. R (via Homebrew or CRAN installer)
+brew install --cask r
+
+# 2. Python 3 + pre-commit
+brew install python
+pip3 install pre-commit
+
+# 3. Docker Desktop
+brew install --cask docker
+# Open Docker Desktop at least once to finish setup
+
+# 4. R development tools (RStudio optional)
+brew install --cask rstudio  # optional
+```
+
+</details>
+
+<details>
+<summary><strong>Ubuntu / Debian</strong></summary>
+
+```bash
+# 1. R (CRAN repo for latest version)
+sudo apt update
+sudo apt install -y software-properties-common dirmngr
+wget -qO- https://cloud.r-project.org/bin/linux/ubuntu/marutter_pubkey.asc | \
+  sudo tee -a /etc/apt/trusted.gpg.d/cran_ubuntu_key.asc
+sudo add-apt-repository "deb https://cloud.r-project.org/bin/linux/ubuntu $(lsb_release -cs)-cran40/"
+sudo apt install -y r-base r-base-dev
+
+# 2. System libraries required by R packages
+sudo apt install -y libcurl4-openssl-dev libssl-dev libxml2-dev \
+  libfontconfig1-dev libharfbuzz-dev libfribidi-dev \
+  libfreetype6-dev libpng-dev libtiff5-dev libjpeg-dev
+
+# 3. Python 3 + pre-commit
+sudo apt install -y python3 python3-pip
+pip3 install pre-commit
+
+# 4. Docker
+curl -fsSL https://get.docker.com | sh
+sudo usermod -aG docker $USER
+# Log out and back in for group membership to take effect
+```
+
+</details>
+
+<details>
+<summary><strong>Windows (WSL2)</strong></summary>
+
+```bash
+# Inside WSL2 (Ubuntu), follow the Ubuntu instructions above.
+# Install Docker Desktop for Windows with WSL2 backend.
+```
+
+</details>
+
+### Quick start
+
+```bash
+# 1. Clone and install everything
+git clone https://github.com/metasurveyr/metasurvey.git
+cd metasurvey
+make dev-setup        # installs R deps + pre-commit hooks
+
+# 2. Verify the package
+make test             # run the 2700+ test suite
+make pre-commit       # run all code quality hooks
+make check            # full R CMD check
+
+# 3. Start the API + MongoDB stack
+cp .env.example .env  # edit with your values (or use defaults for local dev)
+make docker-up        # starts mongo, api, worker, shiny
+make docker-seed      # loads recipes, workflows, users, ANDA metadata
+
+# 4. Verify the stack
+curl http://localhost:8787/health   # API
+curl http://localhost:8788/health   # Worker
+open http://localhost:3838          # Shiny Explorer
+```
+
+Run `make help` to see all available targets.
+
+### Pre-commit hooks
+
+Pre-commit runs automatically on every commit. The hooks enforce:
+
+| Hook | What it does |
+|------|-------------|
+| **R code quality** | |
+| `style-files` | Formats R code with `styler::tidyverse_style()` |
+| `roxygenize` | Regenerates `man/` and `NAMESPACE` |
+| `lintr` | Lints R files against `.lintr` rules |
+| `use-tidy-description` | Keeps `DESCRIPTION` fields sorted |
+| `spell-check` | Spell-checks documentation (uses `inst/WORDLIST`) |
[…]
