### 2026-03-01 b7090a6
feat: redesign shiny auth page
- Translate login/register forms to Spanish
- Add feature list with icons (publica recetas, reutiliza workflows, etc)
- Add social proof stats bar (Open Source, Gratis, R Nativo)
- Add registration CTA with gradient button
- Add switch hint between login/register tabs
- New CSS for auth-stats-bar, auth-legal-note, btn-register-cta
- Adjust branding panel gradient and glow effects
