### 2026-03-01 2a45b2a
feat: add Redis cache layer and async compute queue
Cache:
- GET endpoints cached in Redis with TTL (24h ANDA, 1h indicators, 5min recipes/workflows, 60s lists)
- Automatic invalidation on POST/PUT mutations via SCAN pattern delete
- Graceful degradation: no Redis = no cache, everything works as before

Async compute:
- POST /indicators/compute returns 202 + job_id when async enabled
- Worker consumes jobs via BRPOP loop in background process (callr::r_bg)
- New GET /jobs/<id> endpoint for polling job status
- Falls back to synchronous compute when METASURVEY_ENABLE_ASYNC=0

Infrastructure:
- 
[…]
