+++ inst/api/plumber.R
@@ -98,6 +98,11 @@ ENABLE_INDICATORS <- tolower(Sys.getenv(
 ENABLE_WORKER <- tolower(Sys.getenv(
   "METASURVEY_ENABLE_WORKER", "0"
 )) %in% c("1", "true", "yes")
+REDIS_URL <- Sys.getenv("REDIS_URL", "")
+ENABLE_REDIS <- nzchar(REDIS_URL)
+ENABLE_ASYNC <- tolower(Sys.getenv(
+  "METASURVEY_ENABLE_ASYNC", "0"
+)) %in% c("1", "true", "yes")
 
 if (!nzchar(MONGO_URI)) {
   stop(
@@ -168,6 +173,153 @@ message(sprintf(
   "[metasurvey-api] Features — indicators: %s, worker: %s",
   if (ENABLE_INDICATORS) "ON" else "OFF",
   if (ENABLE_WORKER) "ON" else "OFF"
+))
+
+# -- Redis connection ---------------------------------------------------------
+
+redis_con <- NULL
+
+if (ENABLE_REDIS) {
+  redis_con <- tryCatch(
+    redux::hiredis(url = REDIS_URL),
+    error = function(e) {
+      message(sprintf(
+        "[metasurvey-api] Redis connection failed (disabled): %s",
+        e$message
+      ))
+      NULL
+    }
+  )
+  if (!is.null(redis_con)) {
+    message(sprintf(
+      "[metasurvey-api] Redis connected: %s",
+      sub("://.*@", "://***@", REDIS_URL)
+    ))
+  }
+} else {
+  message("[metasurvey-api] Redis: disabled (REDIS_URL not set)")
+}
+
+# -- Cache helpers ------------------------------------------------------------
+
+cache_params_hash <- function(...) {
+  params <- list(...)
+  params <- params[order(names(params))]
+  digest::digest(params, algo = "xxhash32", serialize = TRUE)
+}
+
+cache_get <- function(key) {
+  if (is.null(redis_con)) {
+    return(NULL)
+  }
+  raw <- tryCatch(redis_con$GET(key), error = function(e) NULL)
+  if (is.null(raw)) {
+    return(NULL)
+  }
+  tryCatch(
+    jsonlite::fromJSON(raw, simplifyVector = FALSE),
+    error = function(e) NULL
+  )
+}
+
+cache_set <- function(key, value, ttl) {
+  if (is.null(redis_con)) {
+    return(invisible(NULL))
+  }
+  json <- tryCatch(
+    jsonlite::toJSON(value, auto_unbox = TRUE, null = "null"),
+    error = function(e) NULL
+  )
+  if (is.null(json)) {
+    return(invisible(NULL))
+  }
+  tryCatch(
+    redis_con$SETEX(key, as.integer(ttl), json),
+    error = function(e) invisible(NULL)
+  )
+}
+
+cache_del <- function(...) {
+  if (is.null(redis_con)) {
+    return(invisible(NULL))
+  }
+  keys <- c(...)
+  for (k in keys) {
+    tryCatch(redis_con$DEL(k), error = function(e) invisible(NULL))
+  }
+}
+
+cache_del_pattern <- function(pattern) {
+  if (is.null(redis_con)) {
+    return(invisible(NULL))
+  }
+  tryCatch(
+    {
+      cursor <- "0"
+      repeat {
+        result <- redis_con$SCAN(cursor, "MATCH", pattern, "COUNT", "100")
+        cursor <- result[[1]]
+        keys <- result[[2]]
+        if (length(keys) > 0) {
+          for (k in keys) redis_con$DEL(k)
+        }
+        if (cursor == "0") break
+      }
+    },
[…]
