+++ inst/worker/plumber.R
@@ -60,6 +60,11 @@ db_workflows <- mongolite::mongo(
   db = DATABASE, url = MONGO_URI
 )
 
+db_indicators <- mongolite::mongo(
+  collection = "indicators",
+  db = DATABASE, url = MONGO_URI
+)
+
 message(sprintf(
   "[metasurvey-worker] Connected to MongoDB (db: %s)",
   DATABASE
@@ -69,6 +74,33 @@ message(sprintf(
   DATA_DIR,
   if (dir.exists(DATA_DIR)) "exists" else "NOT FOUND"
 ))
+
+# -- Redis connection ---------------------------------------------------------
+
+REDIS_URL <- Sys.getenv("REDIS_URL", "")
+
+redis_con <- NULL
+
+if (nzchar(REDIS_URL)) {
+  redis_con <- tryCatch(
+    redux::hiredis(url = REDIS_URL),
+    error = function(e) {
+      message(sprintf(
+        "[metasurvey-worker] Redis connection failed: %s",
+        e$message
+      ))
+      NULL
+    }
+  )
+  if (!is.null(redis_con)) {
+    message(sprintf(
+      "[metasurvey-worker] Redis connected: %s",
+      sub("://.*@", "://***@", REDIS_URL)
+    ))
+  }
+} else {
+  message("[metasurvey-worker] Redis: disabled (REDIS_URL not set)")
+}
 
 # -- Helpers ------------------------------------------------------------------
 
@@ -87,7 +119,9 @@ find_survey_file <- function(survey_type, edition) {
 
   for (p in patterns) {
     path <- file.path(base, p)
-    if (file.exists(path)) return(path)
+    if (file.exists(path)) {
+      return(path)
+    }
   }
 
   NULL
@@ -122,7 +156,9 @@ fetch_recipe <- function(recipe_id) {
     limit = 1
   )
   doc <- iter$one()
-  if (is.null(doc)) return(NULL)
+  if (is.null(doc)) {
+    return(NULL)
+  }
   doc
 }
 
@@ -133,7 +169,9 @@ fetch_workflow <- function(workflow_id) {
     limit = 1
   )
   doc <- iter$one()
-  if (is.null(doc)) return(NULL)
+  if (is.null(doc)) {
+    return(NULL)
+  }
   doc
 }
 
@@ -203,8 +241,8 @@ execute_estimation <- function(svy, calls, estimation_type, filters) {
 
     # If by_variable is specified, wrap non-svyby calls
     if (!is.null(filters$by_variable) &&
-        nzchar(filters$by_variable) &&
-        !grepl("svyby", call_str)) {
+      nzchar(filters$by_variable) &&
+      !grepl("svyby", call_str)) {
       # Transform svymean(~var, ...) -> svyby(~var, ~by, design, svymean, ...)
       call_str <- transform_to_svyby(call_str, filters$by_variable)
     }
@@ -383,6 +421,467 @@ format_estimation <- function(estimation, fn_name, call_str) {
   data.table::rbindlist(rows, fill = TRUE)
 }
 
+# -- Job execution (shared by POST /compute and queue consumer) ---------------
+
+#' Execute a compute job. Returns list(results=...) on success or
+#' list(error=...) on failure.
+execute_job <- function(params, db_rec, db_wf) {
+  survey_type <- params$survey_type
+  edition <- params$edition
+  workflow_id <- params$workflow_id
+  recipe_ids <- params$recipe_ids
+  estimation_type <- params$estimation_type %||% "annual"
+  weight_var <- params$weight
+  filters <- params$filters %||% list()
+
+  # 1. Fetch workflow
+  wf <- fetch_workflow(workflow_id)
[…]
