+++ reference/provenance_to_json.md
@@ -44,7 +44,7 @@ provenance_to_json(prov)
 #> {
 #>   "source": {
 #>     "path": null,
-#>     "timestamp": "2026-02-28T14:59:45",
+#>     "timestamp": "2026-03-01T04:26:44",
 #>     "initial_n": 5,
 #>     "hash": null
 #>   },
