### 2026-03-01 51260f7
perf: migrate Docker images to r2u for pre-compiled R packages
Replace rocker/r-ver:4.4.0 with rocker/r2u:24.04 — all CRAN packages
install as pre-compiled Ubuntu binaries via apt, no compilation needed.
System dependencies resolved automatically by r2u.

Build time drops from ~10 min to ~30 sec.
