### 2026-03-01 a303be1
fix: correct Dockerfile path in railway.toml
Railway root directory is inst/api/, so the path should be
relative to that — just "Dockerfile", not "inst/api/Dockerfile".
