### 2026-03-01 cd11e05
fix: parse Redis URL with authentication for redux
redux::hiredis(url=) doesn't handle redis://user:pass@host:port format.
Parse URL manually and pass host/port/password separately, then AUTH.
