+++ inst/worker/plumber.R
@@ -81,9 +81,34 @@ REDIS_URL <- Sys.getenv("REDIS_URL", "")
 
 redis_con <- NULL
 
+# Parse redis:// URL into host/port/password for redux::hiredis()
+parse_redis_url <- function(url) {
+  parsed <- regmatches(
+    url,
+    regexec("^redis://(?:([^:]*):([^@]*)@)?([^:]+):?(\\d+)?$", url, perl = TRUE)
+  )[[1]]
+  if (length(parsed) == 0) {
+    return(NULL)
+  }
+  list(
+    host = parsed[4],
+    port = as.integer(if (nzchar(parsed[5])) parsed[5] else "6379"),
+    password = if (nzchar(parsed[3])) parsed[3] else NULL
+  )
+}
+
 if (nzchar(REDIS_URL)) {
+  redis_params <- parse_redis_url(REDIS_URL)
   redis_con <- tryCatch(
-    redux::hiredis(url = REDIS_URL),
+    if (!is.null(redis_params) && !is.null(redis_params$password)) {
+      r <- redux::hiredis(host = redis_params$host, port = redis_params$port)
+      r$AUTH(redis_params$password)
+      r
+    } else if (!is.null(redis_params)) {
+      redux::hiredis(host = redis_params$host, port = redis_params$port)
+    } else {
+      redux::hiredis(url = REDIS_URL)
+    },
     error = function(e) {
       message(sprintf(
         "[metasurvey-worker] Redis connection failed: %s",
@@ -532,7 +557,31 @@ run_queue_consumer <- function(redis_url, mongo_uri, db_name, data_dir) {
 
   `%||%` <- function(x, y) if (is.null(x)) y else x
 
-  redis <- redux::hiredis(url = redis_url)
+  # Parse redis:// URL into host/port/password
+  parse_redis_url_local <- function(url) {
+    parsed <- regmatches(
+      url,
+      regexec("^redis://(?:([^:]*):([^@]*)@)?([^:]+):?(\\d+)?$", url, perl = TRUE)
+    )[[1]]
+    if (length(parsed) == 0) {
+      return(NULL)
+    }
+    list(
+      host = parsed[4],
+      port = as.integer(if (nzchar(parsed[5])) parsed[5] else "6379"),
+      password = if (nzchar(parsed[3])) parsed[3] else NULL
+    )
+  }
+
+  rp <- parse_redis_url_local(redis_url)
+  if (!is.null(rp) && !is.null(rp$password)) {
+    redis <- redux::hiredis(host = rp$host, port = rp$port)
+    redis$AUTH(rp$password)
+  } else if (!is.null(rp)) {
+    redis <- redux::hiredis(host = rp$host, port = rp$port)
+  } else {
+    redis <- redux::hiredis(url = redis_url)
+  }
   db_rec <- mongolite::mongo("recipes", db = db_name, url = mongo_uri)
   db_wf <- mongolite::mongo("workflows", db = db_name, url = mongo_uri)
   db_ind <- mongolite::mongo("indicators", db = db_name, url = mongo_uri)
