+++ reference/provenance.md
@@ -53,7 +53,7 @@ svy <- Survey$new(
 )
 provenance(svy)
 #> ── Data Provenance ─────────────────────────────────────────────────────────────
-#> Loaded: 2026-03-01T04:50:34 
+#> Loaded: 2026-03-01T04:56:50 
 #> Initial rows: 10 
 #> 
 #> Environment:
