### 2026-03-01 45bf71d
fix: bypass redux REDIS_URL env var interception
redux 1.1.5 internally reads the REDIS_URL env var and tries to parse
it with hiredis C library even when host/port are passed explicitly.
Authenticated URLs (redis://user:pass@host:port) fail to parse with
hiredis 1.2.0 on Ubuntu 24.04.

- Build clean redis://host:port URL and pass via url= parameter
- Handle AUTH separately after connection
- Support individual REDIS_HOST/PORT/PASSWORD env vars as primary source
- Fix health check PING comparison (redis_status class != character)
