+++ inst/worker/plumber.R
@@ -78,36 +78,69 @@ message(sprintf(
 # -- Redis connection ---------------------------------------------------------
 
 REDIS_URL <- Sys.getenv("REDIS_URL", "")
+REDIS_HOST <- Sys.getenv("REDIS_HOST", "")
+REDIS_PORT <- Sys.getenv("REDIS_PORT", "6379")
+REDIS_PASSWORD <- Sys.getenv("REDIS_PASSWORD", "")
+ENABLE_REDIS_WORKER <- nzchar(REDIS_URL) || nzchar(REDIS_HOST)
 
 redis_con <- NULL
 
-# Parse redis:// URL into host/port/password for redux::hiredis()
-parse_redis_url <- function(url) {
-  parsed <- regmatches(
-    url,
-    regexec("^redis://(?:([^:]*):([^@]*)@)?([^:]+):?(\\d+)?$", url, perl = TRUE)
-  )[[1]]
-  if (length(parsed) == 0) {
+# Resolve Redis params: individual env vars > parse REDIS_URL
+resolve_redis_params <- function() {
+  host <- Sys.getenv("REDIS_HOST", "")
+  port <- Sys.getenv("REDIS_PORT", "6379")
+  pass <- Sys.getenv("REDIS_PASSWORD", "")
+  if (nzchar(host)) {
+    return(list(
+      host = host,
+      port = as.integer(port),
+      password = if (nzchar(pass)) pass else NULL
+    ))
+  }
+  url <- trimws(Sys.getenv("REDIS_URL", ""))
+  if (!grepl("^redis://", url)) {
     return(NULL)
   }
-  list(
-    host = parsed[4],
-    port = as.integer(if (nzchar(parsed[5])) parsed[5] else "6379"),
-    password = if (nzchar(parsed[3])) parsed[3] else NULL
-  )
+  rest <- sub("^redis://", "", url)
+  password <- NULL
+  if (grepl("@", rest)) {
+    auth <- sub("@.*$", "", rest)
+    rest <- sub("^[^@]+@", "", rest)
+    if (grepl(":", auth)) {
+      password <- sub("^[^:]*:", "", auth)
+    } else {
+      password <- auth
+    }
+  }
+  rest <- sub("/.*$", "", rest)
+  parts <- strsplit(rest, ":")[[1]]
+  host <- parts[1]
+  port <- if (length(parts) >= 2) as.integer(parts[2]) else 6379L
+  if (!nzchar(host)) {
+    return(NULL)
+  }
+  list(host = host, port = port, password = password)
 }
 
-if (nzchar(REDIS_URL)) {
-  redis_params <- parse_redis_url(REDIS_URL)
+if (ENABLE_REDIS_WORKER) {
+  redis_params <- resolve_redis_params()
+  message(sprintf(
+    "[metasurvey-worker] Connecting to Redis: %s:%s",
+    if (!is.null(redis_params)) redis_params$host else "?",
+    if (!is.null(redis_params)) redis_params$port else "?"
+  ))
   redis_con <- tryCatch(
-    if (!is.null(redis_params) && !is.null(redis_params$password)) {
-      r <- redux::hiredis(host = redis_params$host, port = redis_params$port)
-      r$AUTH(redis_params$password)
-      r
-    } else if (!is.null(redis_params)) {
-      redux::hiredis(host = redis_params$host, port = redis_params$port)
-    } else {
-      redux::hiredis(url = REDIS_URL)
+    {
+      if (!is.null(redis_params)) {
+        clean_url <- sprintf("redis://%s:%d", redis_params$host, redis_params$port)
+        r <- redux::hiredis(url = clean_url)
+        if (!is.null(redis_params$password)) {
+          r$AUTH(redis_params$password)
+        }
+        r
+      } else {
+        redux::hiredis()
[…]
