+++ reference/provenance.md
@@ -53,7 +53,7 @@ svy <- Survey$new(
 )
 provenance(svy)
 #> ── Data Provenance ─────────────────────────────────────────────────────────────
-#> Loaded: 2026-03-01T04:56:50 
+#> Loaded: 2026-03-01T05:31:42 
 #> Initial rows: 10 
 #> 
 #> Environment:
