### 2026-03-01 cd8f43d
fix: eliminate double Docker build in CI/CD pipeline
GH Actions was building images to GHCR and then railway up was
rebuilding from source on Railway. Now the pipeline is:

1. GH Actions builds + pushes to ghcr.io (single build)
2. Railway pulls the pre-built image (railway redeploy, no rebuild)

Requires one-time Railway dashboard config: change service source
from GitHub Repo to Docker Image pointing to the GHCR image.
