### 2026-03-01 e7d7b43
chore: remove debug field from health endpoint
