+++ reference/read_workflow.md
@@ -41,6 +41,6 @@ wf <- RecipeWorkflow$new(
 )
 f <- tempfile(fileext = ".json")
 save_workflow(wf, f)
-#> Workflow saved to /tmp/RtmpEdkAsK/file1eca72151fd6.json
+#> Workflow saved to /tmp/Rtmp8JVcQQ/file1e9f20cc123a.json
 wf2 <- read_workflow(f)
 ```
