+++ reference/provenance_to_json.md
@@ -44,7 +44,7 @@ provenance_to_json(prov)
 #> {
 #>   "source": {
 #>     "path": null,
-#>     "timestamp": "2026-03-01T05:40:31",
+#>     "timestamp": "2026-03-01T05:51:44",
 #>     "initial_n": 5,
 #>     "hash": null
 #>   },
