+++ articles/workflows-and-estimation.md
@@ -392,7 +392,7 @@ to trace any estimate back to the raw data.
 prov <- provenance(svy_full)
 prov
 #> ── Data Provenance ─────────────────────────────────────────────────────────────
-#> Loaded: 2026-03-01T05:42:21 
+#> Loaded: 2026-03-01T05:53:40 
 #> Initial rows: 200 
 #> 
 #> Environment:
