+++ reference/get_backend.md
@@ -41,6 +41,6 @@ backend
 #>     search: function (query) 
 #>     type: local
 #>   Private:
-#>     .path: /tmp/RtmpEdkAsK/file1eca19980f96.json
+#>     .path: /tmp/Rtmp8JVcQQ/file1e9f4de67f26.json
 #>     .registry: RecipeRegistry, R6
 ```
