### 2026-03-01 70b067b
security: harden API against pentest findings
- Validate input types on /auth/login and /auth/register to prevent
  NoSQL injection via MongoDB operator payloads ($gt, $ne, etc.)
- Add Redis-backed rate limiting on /auth/login (5 attempts/5min per IP)
- Replace specific "Email already registered" with generic error message
  to prevent user enumeration
- Add security headers: X-Content-Type-Options, X-Frame-Options, HSTS,
  CSP, X-XSS-Protection, Referrer-Policy, Permissions-Policy
- Disable Swagger UI by default (METASURVEY_ENABLE_DOCS=1 to enable)
- Strip user_info.email from all public GET 
[…]
