+++ inst/api/plumber.R
@@ -30,6 +30,7 @@
 #   METASURVEY_WORKER_URL          — Internal worker URL (e.g. http://worker:8788)
 #   METASURVEY_ENABLE_INDICATORS   — "1" (default) or "0" to disable /indicators
 #   METASURVEY_ENABLE_WORKER       — "0" (default) or "1" to enable POST /compute
+#   METASURVEY_ENABLE_DOCS         — "0" (default) or "1" to enable Swagger UI
 #
 # Launch:
 #   METASURVEY_MONGO_URI="mongodb+srv://user:pass@cluster.mongodb.net" \
@@ -47,6 +48,13 @@ library(jose)
 #* @plumber
 function(pr) {
   pr$setSerializer(serializer_json(auto_unbox = TRUE))
+
+  # Disable Swagger UI in production (set METASURVEY_ENABLE_DOCS=1 to enable)
+  enable_docs <- tolower(Sys.getenv("METASURVEY_ENABLE_DOCS", "0")) %in%
+    c("1", "true", "yes")
+  if (!enable_docs) {
+    pr$setDocs(FALSE)
+  }
 
   # Enhance OpenAPI spec with security scheme and API info
   spec <- pr$getApiSpec()
@@ -106,6 +114,13 @@ ENABLE_REDIS <- nzchar(REDIS_URL) || nzchar(REDIS_HOST)
 ENABLE_ASYNC <- tolower(Sys.getenv(
   "METASURVEY_ENABLE_ASYNC", "0"
 )) %in% c("1", "true", "yes")
+ENABLE_DOCS <- tolower(Sys.getenv(
+  "METASURVEY_ENABLE_DOCS", "0"
+)) %in% c("1", "true", "yes")
+
+# Rate limiting config
+RATE_LIMIT_MAX <- 5L # max attempts per window
+RATE_LIMIT_WINDOW <- 300L # 5 minutes
 
 if (!nzchar(MONGO_URI)) {
   stop(
@@ -331,6 +346,36 @@ cache_del_pattern <- function(pattern) {
   )
 }
 
+# -- Rate limiting helpers ---------------------------------------------------
+
+rate_limit_check <- function(action, client_ip) {
+  if (is.null(redis_con)) {
+    return(TRUE)
+  }
+  key <- sprintf("ratelimit:%s:%s", action, client_ip)
+  tryCatch(
+    {
+      count <- redis_con$INCR(key)
+      if (identical(as.integer(count), 1L)) {
+        redis_con$EXPIRE(key, RATE_LIMIT_WINDOW)
+      }
+      as.integer(count) <= RATE_LIMIT_MAX
+    },
+    error = function(e) TRUE
+  )
+}
+
+rate_limit_remaining <- function(action, client_ip) {
+  if (is.null(redis_con)) {
+    return(RATE_LIMIT_MAX)
+  }
+  key <- sprintf("ratelimit:%s:%s", action, client_ip)
+  count <- tryCatch(as.integer(redis_con$GET(key) %||% "0"),
+    error = function(e) 0L
+  )
+  max(0L, RATE_LIMIT_MAX - count)
+}
+
 # -- Job queue helpers --------------------------------------------------------
 
 JOB_TTL <- 3600L
@@ -379,8 +424,9 @@ job_enqueue <- function(job_id) {
 }
 
 message(sprintf(
-  "[metasurvey-api] Features — async: %s",
-  if (ENABLE_ASYNC && !is.null(redis_con)) "ON" else "OFF"
+  "[metasurvey-api] Features — async: %s, docs: %s",
+  if (ENABLE_ASYNC && !is.null(redis_con)) "ON" else "OFF",
+  if (ENABLE_DOCS) "ON" else "OFF"
 ))
 
[…]
