### 2026-03-02 91a33c7
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/lorenzwalthert/precommit: v0.4.3 → v0.4.3.9020](https://github.com/lorenzwalthert/precommit/compare/v0.4.3...v0.4.3.9020)
- [github.com/pre-commit/pre-commit-hooks: v5.0.0 → v6.0.0](https://github.com/pre-commit/pre-commit-hooks/compare/v5.0.0...v6.0.0)
