+++ .pre-commit-config.yaml
@@ -8,7 +8,7 @@
 repos:
 # ── R-specific hooks (lorenzwalthert/precommit) ─────────────
 -   repo: https://github.com/lorenzwalthert/precommit
-    rev: v0.4.3
+    rev: v0.4.3.9020
     hooks:
     -   id: style-files
         args: [--style_pkg=styler, --style_fun=tidyverse_style]
@@ -63,7 +63,7 @@ repos:
 
 # ── General file quality hooks ───────────────────────────────
 -   repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v5.0.0
+    rev: v6.0.0
     hooks:
     -   id: check-added-large-files
         args: ['--maxkb=200']
