+++ reference/provenance.md
@@ -53,7 +53,7 @@ svy <- Survey$new(
 )
 provenance(svy)
 #> ── Data Provenance ─────────────────────────────────────────────────────────────
-#> Loaded: 2026-03-01T06:12:13 
+#> Loaded: 2026-03-02T22:21:04 
 #> Initial rows: 10 
 #> 
 #> Environment:
