+++ reference/read_workflow.md
@@ -41,6 +41,6 @@ wf <- RecipeWorkflow$new(
 )
 f <- tempfile(fileext = ".json")
 save_workflow(wf, f)
-#> Workflow saved to /tmp/RtmpnFSMDj/file1e886dc02efd.json
+#> Workflow saved to /tmp/RtmpyMWuZW/file1e981994cf49.json
 wf2 <- read_workflow(f)
 ```
