### 2026-03-12 edc5e1a
fix: replace dontrun with donttest and fix long code lines in R source files
Co-authored-by: mauroloprete <55565968+mauroloprete@users.noreply.github.com>
