### 2026-03-12 0c63819
fix: fix all long code lines (>80 chars) in test files to pass goodpractice linter
Co-authored-by: mauroloprete <55565968+mauroloprete@users.noreply.github.com>
