+++ articles/ech-demographics-recipe-es.md
@@ -355,13 +355,13 @@ cat(readLines(f, n = 15), sep = "\n")
 #>   "description": "Demografia armonizada: sexo, grupo de edad, relacion, nivel educativo y area geografica.",
 #>   "topic": "demographics",
 #>   "doi": {},
-#>   "id": "r_1772490109_663",
+#>   "id": "r_1775514707_663",
 #>   "version": "1.0.0",
 #>   "downloads": 0,
 #>   "categories": [],
 #>   "certification": {
 #>     "level": "community",
-#>     "certified_at": "2026-03-02 22:21:49.411863"
+#>     "certified_at": "2026-04-06 22:31:47.775551"
 ```
 
 ## Aplicando a una nueva edicion
