+++ reference/provenance.md
@@ -53,11 +53,11 @@ svy <- Survey$new(
 )
 provenance(svy)
 #> ── Data Provenance ─────────────────────────────────────────────────────────────
-#> Loaded: 2026-03-02T22:21:04 
+#> Loaded: 2026-04-06T22:30:05 
 #> Initial rows: 10 
 #> 
 #> Environment:
 #>   metasurvey: 0.0.21 
-#>   R: 4.5.2 
+#>   R: 4.5.3 
 #>   survey: 4.5 
 ```
