+++ news/index.md
@@ -722,8 +722,7 @@ CRAN release: 2026-02-25
 ### Bug fixes
 
 - Fixed Codecov GitHub Actions workflow: replaced deprecated
-  [`covr::codecov()`](http://covr.r-lib.org/reference/codecov.md) with
-  [`covr::package_coverage()`](http://covr.r-lib.org/reference/package_coverage.md) +
+  `covr::codecov()` with `covr::package_coverage()` +
   `codecov/codecov-action@v5` for reliable coverage uploads with token
   authentication.
 - Added CI guard to `.Rprofile` to skip example-data setup in GitHub
