### 2026-04-17 560f4b4
feat: add conf.level parameter to workflow()
Allow users to control the confidence level of the interval
(default 0.95). The parameter is passed through to stats::confint()
for all estimation types (svymean, svyby, svyratio, convey).
Fallback z-score calculations also respect the configured level.
