+++ R/workflow.R
@@ -14,6 +14,8 @@
 #'   determines which weight to use. Options:
 #'   `"monthly"`, `"quarterly"`, `"annual"`, or vector
 #'   with multiple types
+#' @param conf.level Confidence level for the interval (default `0.95`).
+#'   Passed to \code{\link[stats]{confint}}.
 #'
 #' @return \code{data.table} with results from all
 #'   estimations, including columns:
@@ -78,16 +80,19 @@
 #' @family workflows
 #' @export
 
-workflow <- function(svy, ..., estimation_type = "monthly") {
+workflow <- function(svy, ..., estimation_type = "monthly",
+                     conf.level = 0.95) {
   if (is(svy, "PoolSurvey")) {
     return(workflow_pool(
       svy, ...,
-      estimation_type = estimation_type
+      estimation_type = estimation_type,
+      conf.level = conf.level
     ))
   } else {
     return(workflow_default(
       svy, ...,
-      estimation_type = estimation_type
+      estimation_type = estimation_type,
+      conf.level = conf.level
     ))
   }
 }
@@ -103,7 +108,8 @@ workflow <- function(svy, ..., estimation_type = "monthly") {
 #' @keywords internal
 #' @noRd
 
-workflow_default <- function(survey, ..., estimation_type = "monthly") {
+workflow_default <- function(survey, ..., estimation_type = "monthly",
+                             conf.level = 0.95) {
   .calls <- substitute(list(...))
 
   result <- rbindlist(
@@ -132,7 +138,8 @@ workflow_default <- function(survey, ..., estimation_type = "monthly") {
                     )
 
                     return(cat_estimation(
-                      estimation, name_function
+                      estimation, name_function,
+                      conf.level = conf.level
                     ))
                   }
                 ),
@@ -180,7 +187,8 @@ workflow_default <- function(survey, ..., estimation_type = "monthly") {
 #' @noRd
 
 
-workflow_pool <- function(survey, ..., estimation_type = "monthly") {
+workflow_pool <- function(survey, ..., estimation_type = "monthly",
+                          conf.level = 0.95) {
   if (grepl(":", estimation_type)) {
     estimation_type_first <- strsplit(estimation_type, ":")[[1]][1]
     estimation_type <- strsplit(estimation_type, ":")[[1]][2]
@@ -230,7 +238,8 @@ workflow_pool <- function(survey, ..., estimation_type = "monthly") {
                       )
                     )
                     return(cat_estimation(
-                      estimation, name_function
+                      estimation, name_function,
+                      conf.level = conf.level
                     ))
                   }
                 )
@@ -286,7 +295,7 @@ workflow_pool <- function(survey, ..., estimation_type = "monthly") {
 }
 
 
-cat_estimation <- function(estimation, call) {
+cat_estimation <- function(estimation, call, conf.level = 0.95) {
   class_estimation <- class(estimation)[1]
 
   if (!class_estimation %in% c("svyby", "svyratio", "cvystat")) {
@@ -300,7 +309,8 @@ cat_estimation <- function(estimation, call) {
     ),
     list(
       estimation,
-      call
+      call,
+      conf.level = conf.level
     )
   )
 }
@@ -312,7 +322,7 @@ cat_estimation <- function(estimation, call) {
 #' @keywords internal
 #' @noRd
 
-cat_estimation.svyby <- function(estimation, call) {
[…]
