+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: metasurvey
 Title: Reproducible Survey Data Processing with Step Pipelines
-Version: 0.0.21
+Version: 0.0.22
 URL: https://metasurveyr.github.io/metasurvey/, https://github.com/metasurveyr/metasurvey
 BugReports: https://github.com/metasurveyr/metasurvey/issues
 Authors@R:
@@ -80,6 +80,6 @@ Config/testthat/edition: 3
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.3.3
-Depends: 
+Depends:
     R (>= 4.1)
 VignetteBuilder: knitr
