### 2026-04-17 ca11a32
feat: improve workflow output with variable, evaluate columns
- Add `variable` column to all cat_estimation functions for clean
  filtering and grouping of results
- Add `evaluate` column with CV quality classification via evaluate_cv()
- svyby results no longer embed group labels in stat column; grouping
  variables (region, sexo, etc.) are now separate columns
- Fix matrix-to-vector coercion for SE/cv in cat_estimation.default
  and cat_estimation.svyratio (prevents column name conflicts in
  rbindlist when mixing estimation types)
