+++ NEWS.md
@@ -4,6 +4,11 @@
 * `workflow()` gains a `conf.level` parameter (default `0.95`) to control the
   confidence level of the interval. Passed through to `stats::confint()` for
   all estimation types (`svymean`, `svyby`, `svyratio`, `convey`).
+* `workflow()` output now includes `variable` and `evaluate` columns across
+  all estimation types. `variable` enables clean filtering/grouping;
+  `evaluate` classifies CV quality via `evaluate_cv()`. `svyby` results no
+  longer embed group labels in the `stat` column — grouping variables
+  (`region`, `sexo`, etc.) appear as separate columns.
 
 # metasurvey 0.0.21
 
