+++ reference/save_recipe.md
@@ -50,5 +50,5 @@ r <- recipe(
 )
 f <- tempfile(fileext = ".json")
 save_recipe(r, f)
-#> The recipe has been saved in /tmp/RtmpIWj1q0/file831f65fecb64.json
+#> The recipe has been saved in /tmp/RtmpraBwBw/file83e012690faa.json
 ```
