### 2026-04-17 062ffbb
docs: update workflow() roxygen with new output columns
Document variable, evaluate, confint_lower, confint_upper columns
in @return. Add svyby and conf.level examples. Remove outdated
estimation_type/survey_edition from column list.
