+++ reference/provenance_to_json.md
@@ -44,7 +44,7 @@ provenance_to_json(prov)
 #> {
 #>   "source": {
 #>     "path": null,
-#>     "timestamp": "2026-04-17T01:35:12",
+#>     "timestamp": "2026-04-17T02:36:37",
 #>     "initial_n": 5,
 #>     "hash": null
 #>   },
