+++ reference/provenance.md
@@ -53,7 +53,7 @@ svy <- Survey$new(
 )
 provenance(svy)
 #> ── Data Provenance ─────────────────────────────────────────────────────────────
-#> Loaded: 2026-04-17T01:35:12 
+#> Loaded: 2026-04-17T02:36:36 
 #> Initial rows: 10 
 #> 
 #> Environment:
