+++ reference/workflow_table.md
@@ -168,6 +168,6 @@ x
 
 Excellent
 
-metasurvey 0.0.22 \| CI: 95% \| 2026-04-17
+metasurvey 0.0.23 \| CI: 95% \| 2026-04-17
 
 \# }
