+++ reference/provenance.md
@@ -53,7 +53,7 @@ svy <- Survey$new(
 )
 provenance(svy)
 #> ── Data Provenance ─────────────────────────────────────────────────────────────
-#> Loaded: 2026-04-17T16:35:39 
+#> Loaded: 2026-04-17T16:35:20 
 #> Initial rows: 10 
 #> 
 #> Environment:
