### 2026-04-19 befa63a
fix: warn when conf.level is passed inside estimation call in workflow()
conf.level inside svymean()/svyby() is silently ignored, leading users
to believe intervals change when they don't. Now emits a warning
guiding them to pass conf.level to workflow() directly.
