+++ NEWS.md
@@ -1,3 +1,10 @@
+# metasurvey 0.0.23
+
+## Bug fixes
+* `workflow()` now warns when `conf.level` is passed inside an estimation call
+  (e.g., `svymean(~x, conf.level = 0.9)`) where it is silently ignored. The
+  warning guides the user to pass `conf.level` to `workflow()` instead.
+
 # metasurvey 0.0.22
 
 ## New features
