+++ reference/transpile_coverage.md
@@ -42,8 +42,8 @@ tf <- tempfile(fileext = ".do")
 writeLines(c("gen x = 1", "replace x = 2 if y == 3", "drop z"), tf)
 transpile_coverage(tf)
 #>                  file                                path total_commands
-#> 1 file1eaf4dd0d8f4.do /tmp/RtmpM2hfgs/file1eaf4dd0d8f4.do              2
-#> 2               TOTAL /tmp/RtmpM2hfgs/file1eaf4dd0d8f4.do              2
+#> 1 file20065f8cbc6d.do /tmp/RtmpxR4ALH/file20065f8cbc6d.do              2
+#> 2               TOTAL /tmp/RtmpxR4ALH/file20065f8cbc6d.do              2
 #>   translated skipped manual_review coverage_pct
 #> 1          2       0             0          100
 #> 2          2       0             0          100
