+++ articles/convey-inequality.md
@@ -326,7 +326,7 @@ workflow_table(
 | Statistic                                  | variable | Estimate | SE    | CI Lower | CI Upper | CV (%) | Quality   |
 | :svygini: gini                             | gini     | 0.48     | 0.033 | 0.42     | 0.55     | 6.9    | Very good |
 | :svyatk: atkinson                          | atkinson | 0.02     | 0.002 | 0.01     | 0.02     | 8.8    | Very good |
-| metasurvey 0.0.23 \| CI: 95% \| 2026-04-17 |          |          |       |          |          |        |           |
+| metasurvey 0.0.23 \| CI: 95% \| 2026-04-19 |          |          |       |          |          |        |           |
 
 ## Provenance
 
@@ -337,7 +337,7 @@ convey estimates computed, and package versions — is available:
 prov <- provenance(results)
 prov
 #> ── Data Provenance ─────────────────────────────────────────────────────────────
-#> Loaded: 2026-04-17T16:35:51 
+#> Loaded: 2026-04-19T20:58:58 
 #> Initial rows: 200 
 #> 
 #> Pipeline:
@@ -345,7 +345,7 @@ prov
 #> 
 #> Estimation:
 #>   Type: annual 
-#>   Timestamp: 2026-04-17T16:35:51 
+#>   Timestamp: 2026-04-19T20:58:58 
 #> 
 #> Environment:
 #>   metasurvey: 0.0.23 
