+++ reference/save_recipe.md
@@ -50,5 +50,5 @@ r <- recipe(
 )
 f <- tempfile(fileext = ".json")
 save_recipe(r, f)
-#> The recipe has been saved in /tmp/RtmpM2hfgs/file1eaf4e4435eb.json
+#> The recipe has been saved in /tmp/RtmpxR4ALH/file200666f21d66.json
 ```
