+++ reference/print.metasurvey_provenance.md
@@ -41,10 +41,10 @@ s <- step_compute(s, age2 = age * 2)
 s <- bake_steps(s)
 print(provenance(s))
 #> ── Data Provenance ─────────────────────────────────────────────────────────────
-#> Loaded: 2026-04-17T16:35:20 
+#> Loaded: 2026-04-19T20:57:59 
 #> 
 #> Pipeline:
-#>   1. step_1 Compute: age2  N=48 [0.0ms]
+#>   1. step_1 Compute: age2  N=48 [1.0ms]
 #> 
 #> Environment:
 #>   metasurvey: 0.0.23 
