+++ articles/workflows-and-estimation-es.md
@@ -395,7 +395,7 @@ permite rastrear cualquier estimacion hasta los datos originales.
 prov <- provenance(svy_full)
 prov
 #> ── Data Provenance ─────────────────────────────────────────────────────────────
-#> Loaded: 2026-04-19T21:00:13 
+#> Loaded: 2026-04-19T21:00:27 
 #> Initial rows: 200 
 #> 
 #> Environment:
