+++ reference/provenance_to_json.md
@@ -44,7 +44,7 @@ provenance_to_json(prov)
 #> {
 #>   "source": {
 #>     "path": null,
-#>     "timestamp": "2026-04-19T20:58:00",
+#>     "timestamp": "2026-04-19T20:57:44",
 #>     "initial_n": 5,
 #>     "hash": null
 #>   },
