### 2026-04-23 7a08214
feat: support conf.level inside individual estimation calls in workflow()
conf.level can now be passed inside svymean()/svyby()/svytotal() calls
and it will be extracted and used for that estimation's confidence
intervals, overriding the workflow() default. This is more natural for
users who expect conf.level to work where they write it.
