+++ R/workflow.R
@@ -16,6 +16,9 @@
 #'   with multiple types
 #' @param conf.level Confidence level for the interval (default `0.95`).
 #'   Passed to \code{\link[stats]{confint}}.
+#'   Can also be set per-call inside the estimation function
+#'   (e.g., `svymean(~x, na.rm = TRUE, conf.level = 0.90)`),
+#'   which overrides the `workflow()` default for that estimation.
 #'
 #' @return \code{data.table} with results from all
 #'   estimations, including columns:
@@ -73,12 +76,19 @@
 #'   estimation_type = "annual"
 #' )
 #'
-#' # Custom confidence level (90%)
+#' # Custom confidence level (90%) for all estimations
 #' result_90 <- workflow(
 #'   svy = list(svy),
 #'   survey::svymean(~x, na.rm = TRUE),
 #'   estimation_type = "annual",
 #'   conf.level = 0.90
+#' )
+#'
+#' # Per-call confidence level (overrides workflow default)
+#' result_mixed <- workflow(
+#'   svy = list(svy),
+#'   survey::svymean(~x, na.rm = TRUE, conf.level = 0.80),
+#'   estimation_type = "annual"
 #' )
 #'
 #' @seealso
@@ -123,7 +133,6 @@ workflow <- function(svy, ..., estimation_type = "monthly",
 workflow_default <- function(survey, ..., estimation_type = "monthly",
                              conf.level = 0.95) {
   .calls <- substitute(list(...))
-  .warn_conf_level_in_calls(.calls)
 
   result <- rbindlist(
     lapply(
@@ -143,6 +152,8 @@ workflow_default <- function(survey, ..., estimation_type = "monthly",
                   function(i) {
                     call <- as.list(.calls[[i]])
                     name_function <- deparse(call[[1]])
+                    extracted <- .extract_conf_level(call, conf.level)
+                    call <- as.list(extracted$call)
                     call[["design"]] <- substitute(design)
                     call <- as.call(call)
                     estimation <- eval(
@@ -152,7 +163,7 @@ workflow_default <- function(survey, ..., estimation_type = "monthly",
 
                     return(cat_estimation(
                       estimation, name_function,
-                      conf.level = conf.level
+                      conf.level = extracted$conf.level
                     ))
                   }
                 ),
@@ -211,7 +222,6 @@ workflow_pool <- function(survey, ..., estimation_type = "monthly",
   }
 
   .calls <- substitute(list(...))
-  .warn_conf_level_in_calls(.calls)
 
   if (all(c("rho", "R") %in% names(.calls))) {
     rho <- eval(.calls[["rho"]])
@@ -243,6 +253,8 @@ workflow_pool <- function(survey, ..., estimation_type = "monthly",
                   function(j) {
                     call <- as.list(.calls[[j]])
                     name_function <- deparse(call[[1]])
+                    extracted <- .extract_conf_level(call, conf.level)
+                    call <- as.list(extracted$call)
                     call[["design"]] <- substitute(design)
                     call <- as.call(call)
                     estimation <- eval(
@@ -253,7 +265,7 @@ workflow_pool <- function(survey, ..., estimation_type = "monthly",
                     )
                     return(cat_estimation(
                       estimation, name_function,
-                      conf.level = conf.level
+                      conf.level = extracted$conf.level
                     ))
                   }
                 )
@@ -309,18 +321,13 @@ workflow_pool <- function(survey, ..., estimation_type = "monthly",
 }
 
 
-.warn_conf_level_in_calls <- function(.calls) {
-  for (i in seq.int(2L, length(.calls))) {
[…]
