+++ tests/testthat/test-workflow.R
@@ -544,24 +544,54 @@ test_that("conf.level changes confidence intervals", {
   expect_true(r90$confint_upper < r95$confint_upper)
 })
 
-test_that("conf.level inside estimation call warns", {
-  svy <- make_test_survey(50)
+test_that("conf.level inside estimation call is used", {
+  svy <- make_test_survey(100)
 
-  expect_warning(
-    workflow(list(svy),
-      survey::svymean(~x, na.rm = TRUE, conf.level = 0.8),
-      estimation_type = "annual"
-    ),
-    "conf.level.*inside.*svymean.*Pass it to workflow"
+  r95 <- workflow(list(svy), survey::svymean(~x, na.rm = TRUE),
+    estimation_type = "annual"
+  )
+  r80 <- workflow(list(svy),
+    survey::svymean(~x, na.rm = TRUE, conf.level = 0.80),
+    estimation_type = "annual"
   )
 
-  expect_warning(
-    workflow(list(svy),
-      survey::svyby(~x, ~region, survey::svymean,
-        na.rm = TRUE, conf.level = 0.8
-      ),
-      estimation_type = "annual"
-    ),
-    "conf.level.*inside.*svyby.*Pass it to workflow"
+  expect_equal(r95$value, r80$value)
+  expect_true(r80$confint_lower > r95$confint_lower)
+  expect_true(r80$confint_upper < r95$confint_upper)
+})
+
+test_that("conf.level inside svyby call is used", {
+  svy <- make_test_survey(100)
+
+  r95 <- workflow(list(svy),
+    survey::svyby(~x, ~region, survey::svymean, na.rm = TRUE),
+    estimation_type = "annual"
   )
+  r80 <- workflow(list(svy),
+    survey::svyby(~x, ~region, survey::svymean,
+      na.rm = TRUE,
+      conf.level = 0.80
+    ),
+    estimation_type = "annual"
+  )
+
+  expect_equal(r95$value, r80$value)
+  expect_true(all(r80$confint_lower > r95$confint_lower))
+  expect_true(all(r80$confint_upper < r95$confint_upper))
+})
+
+test_that("per-call conf.level overrides workflow default", {
+  svy <- make_test_survey(100)
+
+  result <- workflow(list(svy),
+    survey::svymean(~x, na.rm = TRUE, conf.level = 0.80),
+    estimation_type = "annual", conf.level = 0.99
+  )
+  r80 <- workflow(list(svy),
+    survey::svymean(~x, na.rm = TRUE),
+    estimation_type = "annual", conf.level = 0.80
+  )
+
+  expect_equal(result$confint_lower, r80$confint_lower)
+  expect_equal(result$confint_upper, r80$confint_upper)
 })
