+++ reference/provenance.md
@@ -53,7 +53,7 @@ svy <- Survey$new(
 )
 provenance(svy)
 #> ── Data Provenance ─────────────────────────────────────────────────────────────
-#> Loaded: 2026-04-19T20:57:43 
+#> Loaded: 2026-04-23T01:08:55 
 #> Initial rows: 10 
 #> 
 #> Environment:
