+++ news/index.md
@@ -2,14 +2,15 @@
 
 ## metasurvey 0.0.23
 
-### Bug fixes
+### New features
 
 - [`workflow()`](https://metasurveyr.github.io/metasurvey/reference/workflow.md)
-  now warns when `conf.level` is passed inside an estimation call (e.g.,
-  `svymean(~x, conf.level = 0.9)`) where it is silently ignored. The
-  warning guides the user to pass `conf.level` to
+  now supports `conf.level` inside individual estimation calls (e.g.,
+  `svymean(~x, na.rm = TRUE, conf.level = 0.80)`). The per-call value
+  overrides the
   [`workflow()`](https://metasurveyr.github.io/metasurvey/reference/workflow.md)
-  instead.
+  default, allowing different confidence levels for each estimation in
+  the same call.
 
 ## metasurvey 0.0.22
 
