+++ reference/provenance_to_json.md
@@ -44,7 +44,7 @@ provenance_to_json(prov)
 #> {
 #>   "source": {
 #>     "path": null,
-#>     "timestamp": "2026-04-23T01:08:56",
+#>     "timestamp": "2026-04-23T01:09:34",
 #>     "initial_n": 5,
 #>     "hash": null
 #>   },
