### 2026-04-23 77db0e0
chore: prepare 0.0.23 for CRAN submission
- Add missing entries to .Rbuildignore (compatibilizada, postulacion,
  ddt_00_2024, linkedin files, convey-inequality_files)
- Fix harmonize() example: use Recipe$new() with correct params, wrap
  in \donttest{} with tempfile-based local backend
- Keep load_survey_example() as \dontrun{} (GitHub data repo not available)
- Add on.exit() backend restoration in all test-harmonize.R tests
- Fix is_baked test: ensure lazy_processing=TRUE to avoid state leakage
- Update cran-comments.md for 0.0.23
