+++ tests/testthat/test-harmonize.R
@@ -1,0 +1,294 @@
+# --- topo_sort_recipes tests ---
+
+test_that("topo_sort_recipes: empty list returns empty", {
+  expect_equal(topo_sort_recipes(list()), list())
+})
+
+test_that("topo_sort_recipes: single recipe returns as-is", {
+  r <- make_eco_recipe("A", "u")
+  result <- topo_sort_recipes(list(r))
+  expect_equal(length(result), 1)
+  expect_equal(result[[1]]$name, "A")
+})
+
+test_that("topo_sort_recipes: respects dependency order", {
+  r1 <- make_eco_recipe("Base", "u")
+  r2 <- make_eco_recipe("Derived", "u")
+  r2$depends_on_recipes <- list(as.character(r1$id))
+
+  result <- topo_sort_recipes(list(r2, r1))
+  expect_equal(result[[1]]$name, "Base")
+  expect_equal(result[[2]]$name, "Derived")
+})
+
+test_that("topo_sort_recipes: diamond dependency", {
+  #    A
+  #   / \
+  #  B   C
+  #   \ /
+  #    D
+  rA <- make_eco_recipe("A", "u")
+  rB <- make_eco_recipe("B", "u")
+  rC <- make_eco_recipe("C", "u")
+  rD <- make_eco_recipe("D", "u")
+  rB$depends_on_recipes <- list(as.character(rA$id))
+  rC$depends_on_recipes <- list(as.character(rA$id))
+  rD$depends_on_recipes <- list(as.character(rB$id), as.character(rC$id))
+
+  result <- topo_sort_recipes(list(rD, rC, rB, rA))
+  names_order <- vapply(result, function(r) r$name, character(1))
+  # A must come before B and C; B and C must come before D
+  expect_true(which(names_order == "A") < which(names_order == "B"))
+  expect_true(which(names_order == "A") < which(names_order == "C"))
+  expect_true(which(names_order == "B") < which(names_order == "D"))
+  expect_true(which(names_order == "C") < which(names_order == "D"))
+})
+
+test_that("topo_sort_recipes: cycle detection", {
+  r1 <- make_eco_recipe("X", "u")
+  r2 <- make_eco_recipe("Y", "u")
+  r1$depends_on_recipes <- list(as.character(r2$id))
+  r2$depends_on_recipes <- list(as.character(r1$id))
+
+  expect_error(
+    topo_sort_recipes(list(r1, r2)),
+    "Cycle detected"
+  )
+})
+
+test_that("topo_sort_recipes: external deps are ignored", {
+  r1 <- make_eco_recipe("A", "u")
+  r1$depends_on_recipes <- list("nonexistent_id_123")
+
+  result <- topo_sort_recipes(list(r1))
+  expect_equal(length(result), 1)
+  expect_equal(result[[1]]$name, "A")
+})
+
+# --- harmonize tests ---
+
+test_that("harmonize: rejects non-list input", {
+  expect_error(harmonize(NULL), "non-empty list")
+  expect_error(harmonize(list()), "non-empty list")
+  expect_error(harmonize("not a list"), "non-empty list")
+})
+
+test_that("harmonize: rejects non-Survey elements", {
+  expect_error(
+    harmonize(list("not a survey")),
+    "not a Survey object"
+  )
+})
+
+test_that("harmonize: rejects invalid grouping", {
+  svy <- make_test_survey(10)
+  expect_error(
+    harmonize(list(svy), grouping = "invalid"),
+    "grouping"
+  )
+})
+
+test_that("harmonize: no recipes warns and returns unchanged", {
+  svy <- make_test_survey(10)
+  old_backend <- getOption("metasurvey.backend")
[…]
