+++ man/harmonize.Rd
@@ -1,0 +1,114 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/harmonize.R
+\name{harmonize}
+\alias{harmonize}
+\title{Harmonize surveys using registry recipes}
+\usage{
+harmonize(
+  surveys,
+  survey_type = NULL,
+  topic = NULL,
+  category = NULL,
+  certification_level = NULL,
+  grouping = "annual",
+  group_name = "series",
+  .verbose = getOption("metasurvey.verbose", TRUE)
+)
+}
+\arguments{
+\item{surveys}{List of \code{\link{Survey}} objects to harmonize.}
+
+\item{survey_type}{Character. Override survey type for recipe lookup.
+If \code{NULL} (default), uses each survey's own \code{type} field.}
+
+\item{topic}{Character. Filter recipes by topic
+(e.g. \code{"compatibilizada"}). Default \code{NULL} (no filter).}
+
+\item{category}{Character. Optional category filter for recipes.}
+
+\item{certification_level}{Character. Optional certification level
+filter (e.g. \code{"official"}, \code{"reviewed"}).}
+
+\item{grouping}{Character. Time hierarchy for the PoolSurvey structure.
+One of \code{"annual"} (default), \code{"quarterly"},
+\code{"monthly"}, \code{"biannual"}.}
+
+\item{group_name}{Character. Name of the group in the PoolSurvey
+structure (default \code{"series"}).}
+
+\item{.verbose}{Logical. Print progress messages (default uses
+\code{metasurvey.verbose} option).}
+}
+\value{
+A \code{\link{PoolSurvey}} object containing the harmonized
+surveys, structured as
+\code{list(<grouping> = list(<group_name> = list(svy1, svy2, ...)))}.
+}
+\description{
+For each survey in the list, fetches matching recipes from the active
+backend (by survey type and edition), sorts them by dependency order,
+applies them via \code{\link{bake_recipes}}, and assembles the results
+into a \code{\link{PoolSurvey}} ready for \code{\link{workflow}}.
+}
+\details{
+The function performs the following steps for each survey:
+\enumerate{
+\item Calls \code{\link{filter_recipes}} with the survey's type and
+edition (plus optional category/certification filters).
+\item Sorts matching recipes using topological sort on
+\code{depends_on_recipes}.
+\item Attaches recipes via \code{\link[=Survey]{add_recipe}} and
+applies them with \code{\link{bake_recipes}}.
+}
+
+If no recipes are found for a survey, it is included in the pool
+unchanged (with a warning).
+}
+\examples{
+\donttest{
+# Set up a local backend with a temp file
+old <- set_backend("local", path = tempfile(fileext = ".json"))
+
+# Create a small survey and a recipe
+dt <- data.table::data.table(
+  id = 1:10, x = 1:10, w = rep(1, 10)
+)
+svy <- Survey$new(
+  data = dt, edition = "2023", type = "test",
+  psu = NULL, engine = "data.table",
+  weight = add_weight(annual = "w")
+)
+r <- Recipe$new(
+  name = "demo", edition = "2023", survey_type = "test",
+  default_engine = "data.table", depends_on = list(),
+  user = "demo", description = "demo recipe",
+  steps = list("step_compute(., z = x * 2)"),
+  id = "demo_recipe"
+)
+publish_recipe(r)
+
+pool <- harmonize(list(svy), .verbose = FALSE)
+
+# Restore backend
+options(metasurvey.backend = old)
+}
+
+}
+\seealso{
[…]
