### 2026-04-23 4651357
feat: add topic filter and vector edition support to recipe registry
- RecipeRegistry$filter() and RecipeBackend$filter() gain topic param
- Edition filter now matches vector editions (e.g., c("2006","2007"))
- Move make_eco_recipe() helper to helper-survey.R for reuse
- Add tests for topic filter and vector edition matching
