### 2026-04-23 1186c76
fix: handle duplicate target names and missing renames from STATA transpiler
- step_compute: detect duplicate LHS names (STATA gen/replace pattern)
  and apply sequentially instead of batch data.table :=
- step_rename: silently skip missing columns instead of erroring,
  matching STATA 'cap rename' behavior for transpiled recipes
