+++ tests/testthat/test-steps-basic.R
@@ -1,6 +1,10 @@
 test_that("step_remove removes columns and records step", {
   df <- data.frame(id = 1:3, a = 1, b = 2, w = 1)
-  s <- Survey$new(data = data.table::data.table(df), edition = "2023", type = "ech", psu = NULL, engine = "data.table", weight = add_weight(annual = "w"))
+  s <- Survey$new(
+    data = data.table::data.table(df), edition = "2023",
+    type = "ech", psu = NULL, engine = "data.table",
+    weight = add_weight(annual = "w")
+  )
 
   s2 <- step_remove(s, a, b)
   expect_true(any(grepl("Remove:", names(s2$steps))))
@@ -11,7 +15,11 @@ test_that("step_remove removes columns and records step", {
 
 test_that("step_rename renames columns and records step", {
   df <- data.frame(id = 1:3, a = 1, w = 1)
-  s <- Survey$new(data = data.table::data.table(df), edition = "2023", type = "ech", psu = NULL, engine = "data.table", weight = add_weight(annual = "w"))
+  s <- Survey$new(
+    data = data.table::data.table(df), edition = "2023",
+    type = "ech", psu = NULL, engine = "data.table",
+    weight = add_weight(annual = "w")
+  )
 
   s2 <- step_rename(s, alpha = a)
   expect_true(any(grepl("Rename:", names(s2$steps))))
@@ -23,7 +31,11 @@ test_that("step_rename renames columns and records step", {
 test_that("step_join with data.frame RHS performs left/inner/full joins", {
   lhs <- data.frame(id = 1:3, a = c("x", "y", "z"), w = 1)
   rhs <- data.frame(id = c(2, 3, 4), b = c(20, 30, 40))
-  s <- Survey$new(data = data.table::data.table(lhs), edition = "2023", type = "ech", psu = NULL, engine = "data.table", weight = add_weight(annual = "w"))
+  s <- Survey$new(
+    data = data.table::data.table(lhs), edition = "2023",
+    type = "ech", psu = NULL, engine = "data.table",
+    weight = add_weight(annual = "w")
+  )
 
   # Left join keeps all LHS rows
   s_left <- bake_steps(step_join(s, rhs, by = "id", type = "left"))
@@ -42,8 +54,16 @@ test_that("step_join with data.frame RHS performs left/inner/full joins", {
 test_that("step_join accepts Survey as RHS and handles key mapping", {
   lhs <- data.frame(id = 1:2, a = 1:2, w = 1)
   rhs <- data.frame(code = 2:3, b = 10:11, w2 = 1)
-  s_left <- Survey$new(data = data.table::data.table(lhs), edition = "2023", type = "ech", psu = NULL, engine = "data.table", weight = add_weight(annual = "w"))
-  s_right <- Survey$new(data = data.table::data.table(rhs), edition = "2023", type = "ech", psu = NULL, engine = "data.table", weight = add_weight(annual = "w2"))
+  s_left <- Survey$new(
+    data = data.table::data.table(lhs), edition = "2023",
+    type = "ech", psu = NULL, engine = "data.table",
+    weight = add_weight(annual = "w")
+  )
+  s_right <- Survey$new(
+    data = data.table::data.table(rhs), edition = "2023",
+    type = "ech", psu = NULL, engine = "data.table",
+    weight = add_weight(annual = "w2")
+  )
 
   s_out <- bake_steps(step_join(s_left, s_right, by = c("id" = "code"), type = "left"))
   expect_true("b" %in% names(s_out$data))
@@ -317,12 +337,12 @@ test_that("step_rename with mapping parameter works", {
   expect_true("earnings" %in% names(get_data(s2)))
 })
 
[…]
