### 2026-04-23 ea14d6b
feat: add compatibilizada category to shiny explorer + update man pages
- Add "Compatibilizada" filter option and CSS tag in mod_explore/ui_components
- Fix indentation lint issues in ui_components.R
- Update NAMESPACE and man/*.Rd for harmonize() cross-references
- Add WORDLIST entries for new terms
