+++ man/publish_recipe.Rd
@@ -33,6 +33,7 @@ Other recipes:
 \code{\link{bake_recipes}()},
 \code{\link{explore_recipes}()},
 \code{\link{get_recipe}()},
+\code{\link{harmonize}()},
 \code{\link{print.Recipe}()},
 \code{\link{read_recipe}()},
 \code{\link{recipe}()},
