+++ man/explore_recipes.Rd
@@ -40,6 +40,7 @@ Other recipes:
 \code{\link{add_recipe}()},
 \code{\link{bake_recipes}()},
 \code{\link{get_recipe}()},
+\code{\link{harmonize}()},
 \code{\link{print.Recipe}()},
 \code{\link{publish_recipe}()},
 \code{\link{read_recipe}()},
