+++ man/steps_to_recipe.Rd
@@ -62,6 +62,7 @@ Other recipes:
 \code{\link{bake_recipes}()},
 \code{\link{explore_recipes}()},
 \code{\link{get_recipe}()},
+\code{\link{harmonize}()},
 \code{\link{print.Recipe}()},
 \code{\link{publish_recipe}()},
 \code{\link{read_recipe}()},
