### 2026-04-23 7401cf8
chore: update inst/scripts and add dev deps to Suggests
- Update seed_ech_recipes.R with topic enum schema validator
- Update setup_mongodb.js with compatibilizada topic
- Add validate_harmonize.R for recipe validation against IECON .dta files
- Update transpile_iecon_all.R with improved coverage reporting
- Replace print() with cat() in validate_harmonize.R
- Add devtools and mongolite to Suggests (used by inst/scripts)
