### 2026-05-03 #752 @maurolepore
Following up on @emilyriederer's note from March, the editorial team is still working on the right review structure for a package of this size. I'll continue that work and follow up here once we have a concrete next step.

Thanks for your patience and for sharing your work with the rOpenSci community.
