### 2026-07-08 7708e54
feat: metasurvey meta-package (tidyverse-style umbrella) 0.3.0
Turns the metasurvey package into a meta-package for the modularized ecosystem.
A single library(metasurvey) loads and re-exports metasurvey.core,
metasurvey.fromstata, metasurvey.anda, and metasurvey.explorer.backend (128
symbols). The monolith moved to metasurvey-legacy per the rOpenSci review's
modularization suggestion.

- R/reexports*.R: re-export assignments + @rawNamespace directives + doc stub,
  all generated by data-raw/generate_reexports.R.
- R/reproduce.R: high-level orchestration that spans packages (cut C3):
  reprodu
[…]
