+++ R/reexports-namespace.R
@@ -1,0 +1,260 @@
+# GENERADO por data-raw/generate_reexports.R: no editar a mano.
+
+#' @rawNamespace
+#' importFrom(metasurvey.core,add_category)
+#' export(add_category)
+#' importFrom(metasurvey.core,add_recipe)
+#' export(add_recipe)
+#' importFrom(metasurvey.core,add_replicate)
+#' export(add_replicate)
+#' importFrom(metasurvey.core,add_weight)
+#' export(add_weight)
+#' importFrom(metasurvey.core,bake_recipes)
+#' export(bake_recipes)
+#' importFrom(metasurvey.core,bake_steps)
+#' export(bake_steps)
+#' importFrom(metasurvey.core,cat_design)
+#' export(cat_design)
+#' importFrom(metasurvey.core,cat_design_type)
+#' export(cat_design_type)
+#' importFrom(metasurvey.core,certify_recipe)
+#' export(certify_recipe)
+#' importFrom(metasurvey.core,default_categories)
+#' export(default_categories)
+#' importFrom(metasurvey.core,evaluate_cv)
+#' export(evaluate_cv)
+#' importFrom(metasurvey.core,extract_surveys)
+#' export(extract_surveys)
+#' importFrom(metasurvey.core,extract_time_pattern)
+#' export(extract_time_pattern)
+#' importFrom(metasurvey.core,filter_recipes)
+#' export(filter_recipes)
+#' importFrom(metasurvey.core,filter_workflows)
+#' export(filter_workflows)
+#' importFrom(metasurvey.core,find_workflows_for_recipe)
+#' export(find_workflows_for_recipe)
+#' importFrom(metasurvey.core,get_backend)
+#' export(get_backend)
+#' importFrom(metasurvey.core,get_data)
+#' export(get_data)
+#' importFrom(metasurvey.core,get_engine)
+#' export(get_engine)
+#' importFrom(metasurvey.core,get_follow_up)
+#' export(get_follow_up)
+#' importFrom(metasurvey.core,get_implantation)
+#' export(get_implantation)
+#' importFrom(metasurvey.core,get_metadata)
+#' export(get_metadata)
+#' importFrom(metasurvey.core,get_recipe)
+#' export(get_recipe)
+#' importFrom(metasurvey.core,get_steps)
+#' export(get_steps)
+#' importFrom(metasurvey.core,get_workflow_backend)
+#' export(get_workflow_backend)
+#' importFrom(metasurvey.core,group_dates)
+#' export(group_dates)
+#' importFrom(metasurvey.core,harmonize)
+#' export(harmonize)
+#' importFrom(metasurvey.core,has_design)
+#' export(has_design)
+#' importFrom(metasurvey.core,has_recipes)
+#' export(has_recipes)
+#' importFrom(metasurvey.core,has_steps)
+#' export(has_steps)
+#' importFrom(metasurvey.core,is_baked)
+#' export(is_baked)
+#' importFrom(metasurvey.core,lazy_default)
+#' export(lazy_default)
+#' importFrom(metasurvey.core,list_recipes)
+#' export(list_recipes)
+#' importFrom(metasurvey.core,list_workflows)
+#' export(list_workflows)
+#' importFrom(metasurvey.core,load_panel_survey)
+#' export(load_panel_survey)
+#' importFrom(metasurvey.core,load_survey)
+#' export(load_survey)
+#' importFrom(metasurvey.core,load_survey_example)
+#' export(load_survey_example)
+#' importFrom(metasurvey.core,PoolSurvey)
+#' export(PoolSurvey)
+#' importFrom(metasurvey.core,provenance)
+#' export(provenance)
+#' importFrom(metasurvey.core,provenance_diff)
+#' export(provenance_diff)
[…]
