+++ pkgdown.yml
@@ -1,0 +1,6 @@
+pandoc: 3.8.3
+pkgdown: 2.2.1
+pkgdown_sha: ~
+articles:
+  getting-started: getting-started.html
+last_built: 2026-07-08T16:06Z
